### 2025-11-18 c55962a
Move OBZ data to repo instead of downloading from Zenodo (#1427)
To avoid relying on Zenodo, we move the data to the repo itself.

Closes #1404
