### 2025-11-19 ba62bb8
[JOSS] Remove 'https://orcid.org' prefix from ORCIDs in paper.md
