### 2025-11-19 25dc76c
Update Bestie's full guide url
