### 2025-11-19 acaee90
Fix DOIs and ORCIDs for JOSS paper (#1429)
* [JOSS] Remove 'https://doi.org' prefix from DOI's in paper.bib

* [JOSS] Remove 'https://orcid.org' prefix from ORCIDs in paper.md

* Update Bestie's full guide url
