+++ paper/paper.bib
@@ -54,7 +54,7 @@
   pages         = {126267},
   year          = {2025},
   issn          = {0306-2619},
-  doi           = {https://doi.org/10.1016/j.apenergy.2025.126267},
+  doi           = {10.1016/j.apenergy.2025.126267},
   url           = {https://www.sciencedirect.com/science/article/pii/S0306261925009973},
   author        = {Zhi Gao and Matteo Gazzani and Diego A. Tejada-Arango and Abel {Soares Siqueira} and Ni Wang and Madeleine Gibescu and Germ\'{a}n Morales-Espa\~{n}a},
   keywords      = {Energy system optimization, Temporal aggregation, Computational efficiency},
@@ -67,7 +67,7 @@
   pages = {110674},
   year = {2025},
   issn = {0142-0615},
-  doi = {https://doi.org/10.1016/j.ijepes.2025.110674},
+  doi = {10.1016/j.ijepes.2025.110674},
   url = {https://www.sciencedirect.com/science/article/pii/S014206152500225X},
   author = {Diego A. Tejada-Arango and Juha Kiviluoma and Germán Morales-España},
   keywords = {Energy sector coupling, Optimisation modelling, Energy system models, Linear programming (LP), Computational efficiency},
@@ -164,7 +164,7 @@
   pages         = {100902},
   year          = {2022},
   issn          = {2211-467X},
-  doi           = {https://doi.org/10.1016/j.esr.2022.100902},
+  doi           = {10.1016/j.esr.2022.100902},
   url           = {https://www.sciencedirect.com/science/article/pii/S2211467X22000955},
   author        = {Maren Ihlemann and Iasonas Kouveliotis-Lysikatos and Jiangyi Huang and Joseph Dillon and Ciara O'Dwyer and Topi Rasku and Manuel Marin and Kris Poncelet and Juha Kiviluoma},
   keywords      = {Open source tool, Energy system modelling, Energy system analysis, Integrated energy systems, Investment planning, Sector coupling},
@@ -212,7 +212,7 @@
 }
 @article{Sienna,
 author = {Lara, José Daniel and Barrows, Clayton and Thom, Daniel and Krishnamurthy, Dheepak and Callaway, Duncan},
-doi = {https://doi.org/10.1016/j.softx.2021.100747},
+doi = {10.1016/j.softx.2021.100747},
 journal = {SoftwareX},
 month = jul,
 title = {{PowerSystems.jl — A power system data management package for large scale modeling}},
