### 2025-11-19 a0835f1
[JOSS] Multiple affiliation must be quoted (#1430)
