### 2025-11-19 e34871b
Rename column and file group to investment_group (#1425)
Change the column `group` on table `asset` to `investment_group`.
Changes the constraint file group.jl to investment_group.jl.

Closes #1252
