+++ previews/PR1410/10-tutorials/user-input-files/inputs-description.toml
@@ -1,98 +1,0 @@
-# Tulipa Energy Model Basic User Input Descriptions
-
-[asset]
-description = "Unique identifier with the name of the asset."
-
-[capacity]
-description = "Capacity for one unit of the asset (is therefore multiplied by number of existing plus number of invested assets)."
-unit_of_measure = "GW"
-
-[capacity_storage_energy]
-description = "Capacity of one storage unit."
-unit_of_measure = "GWh"
-
-[carrier]
-description = "Energy carrier, for results visualization"
-
-[commission_year]
-description = "Year of commissioning."
-
-[efficiency]
-description = "Efficiency of transfer in `from_asset` -> `to_asset` direction"
-unit_of_measure = "ratio"
-
-[from_asset]
-description = "Name of the asset. Same as the one in the `asset` table."
-
-[initial_export_units]
-description = "Number of existing units in `from_asset` -> `to_asset` direction"
-unit_of_measure = "number"
-
-[initial_import_units]
-description = "Number of existing units in `to_asset` -> `from_asset` direction"
-unit_of_measure = "number"
-
-[initial_storage_units]
-description = "Number of existing storage units"
-unit_of_measure = "number"
-
-[initial_units]
-description = "Number of existing units"
-unit_of_measure = "number"
-
-[is_seasonal]
-description = "Whether seasonal storage (e.g., hydro) or not (e.g., battery)"
-type = "BOOLEAN"
-
-[is_transport]
-description = "Whether a transport flow or not. Transport assets can have flows in both directions and can be invested in."
-type = "BOOLEAN"
-
-[lat]
-description = "Latitude"
-unit_of_measure = "degrees"
-
-[length]
-description = "How many hours in a year, e.g., 8760"
-
-[lon]
-description = "Longitude"
-unit_of_measure = "degrees"
-
-[partition]
-description = "Partition or temporal resolution in the representative periods. For example, for a `uniform` specification `1` is hourly, `2` is every two hours."
-
-[peak_demand]
-description = "Value that multiplies the demand profile time series."
-unit_of_measure = "GW"
-
-[profile_name]
-description = "Name of profile, used to determine data inside the DuckDB table"
-
-[profile_type]
-description = "Type of profile, used to determine DuckDB table with source profile"
-
-[profile_value]
-description = "value of the profile."
-unit_of_measure = "p.u."
-
-[storage_inflows]
-description = "Value that multiplies the inflow profile time series."
-unit_of_measure = "GWh"
-
-[technology]
-description = "Technology of the asset for visualization and aggregation of results"
-
-[to_asset]
-description = "Name of the asset. Same as the one in the `asset` table."
-
-[type]
-description = "Type of energy asset."
-
-[variable_cost]
-description = "Variable cost for the flow."
-unit_of_measure = "EUR/MWh"
-
-[year]
-description = "Milestone year."
-type = "INTEGER"
