### 2025-11-20 1a226ce
Add producer_efficiency to flow_commission table
