### 2025-11-20 5eb9ff8
Move commodity_price from asset_milestone table to flow_milestone table
