### 2025-11-20 eb048c4
Move commodity_price from asset_milestone table to flow_milestone table
