+++ utils/scripts/delete-a-column.jl
@@ -13,7 +13,7 @@ for dir in dirs
     println("Processing directory: $dir")
 
     # Explicit name of the table
-    filename = joinpath(dir, "profiles-rep-periods.csv")
+    filename = joinpath(dir, "asset-milestone.csv")
 
     connection = DBInterface.connect(DuckDB.DB)
     _q(s) = DuckDB.query(connection, s)
@@ -21,7 +21,7 @@ for dir in dirs
     _q("CREATE TABLE t AS FROM read_csv('$filename')")
 
     # Delete a column by name
-    _q("ALTER TABLE t DROP COLUMN scenario")
+    _q("ALTER TABLE t DROP COLUMN commodity_price")
 
     _q("COPY t TO '$filename' (HEADER, DELIMITER ',')")
 end
