### 2025-11-20 72008a5
Change commodity_price in obj
