### 2025-11-20 c9f9394
Change commodity_price in obj
