### 2025-11-20 d45dc56
Change efficiency to producer_efficiency in obj
