+++ src/objective.jl
@@ -532,10 +532,10 @@ function _create_objective_auxiliary_table(connection, constants)
             flow_commission.investment_cost,
             flow.capacity,
             flow_milestone.commodity_price,
-            asset_commission.efficiency,
+            flow_commission.producer_efficiency,
             flow_milestone.operational_cost,
             -- computed
-            (flow_milestone.commodity_price / asset_commission.efficiency) AS fuel_cost,
+            (flow_milestone.commodity_price / flow_commission.producer_efficiency) AS fuel_cost,
             (fuel_cost + flow_milestone.operational_cost) AS total_variable_cost,
             CASE
                 -- the below closed-form equation does not accept 0 in the denominator when flow.discount_rate = 0
@@ -574,17 +574,6 @@ function _create_objective_auxiliary_table(connection, constants)
         LEFT JOIN flow
             ON flow.from_asset = flow_commission.from_asset
             AND flow.to_asset = flow_commission.to_asset
-        /*
-        The below join works for compact/simple/none method.
-        Note normally this condition milestone_year = commission_year does not work for compact method.
-        But here, it means if you use compact method, the fuel cost will ignore the efficiencies where
-        milestone_year != commission_year
-        It makes sense because if you would like to consider efficiencies from different commission years,
-        you should use semi-compact method instead.
-        */
-        LEFT JOIN asset_commission
-            ON flow_milestone.from_asset = asset_commission.asset
-            AND flow_milestone.milestone_year = asset_commission.commission_year
         ",
     )
 
@@ -599,10 +588,10 @@ function _create_objective_auxiliary_table(connection, constants)
             var.commission_year,
             -- copied over
             flow_milestone.commodity_price,
-            asset_commission.efficiency,
+            flow_commission.producer_efficiency,
             flow_milestone.operational_cost,
             -- computed
-            (flow_milestone.commodity_price / asset_commission.efficiency) AS fuel_cost,
+            (flow_milestone.commodity_price / flow_commission.producer_efficiency) AS fuel_cost,
             (fuel_cost + flow_milestone.operational_cost) AS total_variable_cost,
             CASE
                 -- the below closed-form equation does not accept 0 in the denominator when flow.discount_rate = 0
@@ -630,9 +619,6 @@ function _create_objective_auxiliary_table(connection, constants)
             investment_year_discount * (1 - salvage_value / flow_commission.investment_cost) AS weight_for_flow_investment_discount,
             in_between_years.discount_factor_from_current_milestone_year_to_next_milestone_year AS weight_for_operation_discounts,
         FROM var_vintage_flow AS var
-        LEFT JOIN asset_commission
-            ON var.from_asset = asset_commission.asset
-            AND var.commission_year = asset_commission.commission_year
         LEFT JOIN flow_milestone
             ON var.from_asset = flow_milestone.from_asset
             AND var.to_asset = flow_milestone.to_asset
