+++ docs/src/40-scientific-foundation/40-formulation.md
@@ -103,7 +103,6 @@ In addition, the following subsets represent methods for incorporating additiona
 | Name                                                 | Domain                   | Domains of Indices                                                                                 | Description                                                                                                            | Units                     |
 | ---------------------------------------------------- | ------------------------ | -------------------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------- | ------------------------- |
 | $p^{\text{inv cost}}_{a,y}$                          | $\mathbb{R}_{+}$         | $a \in \mathcal{A}$, $y \in \mathcal{Y}$                                                           | Overnight cost of a unit of asset $a$ at year $y$                                                                      | [kEUR/MW]                 |
-| $p^{\text{commodity price}}_{a,y}$                   | $\mathbb{R}_{+}$         | $a \in \mathcal{A}$, $y \in \mathcal{Y}$                                                           | Price of the commodity to produce a unit of energy for asset $a$ at year $y$. Only applies to producers.               | [kEUR/MWh] or other units |
 | $p^{\text{annualized inv cost}}_{a,y}$               | $\mathbb{R}_{+}$         | $a \in \mathcal{A}$, $y \in \mathcal{Y}$                                                           | Annualized investment cost of a unit of asset $a$ at year $y$                                                          | [kEUR/MW/year]            |
 | $p^{\text{salvage value}}_{a,y}$                     | $\mathbb{R}_{+}$         | $a \in \mathcal{A}$, $y \in \mathcal{Y}$                                                           | Salvage value of a unit of asset $a$ at year $y$                                                                       | [kEUR/MW]                 |
 | $p^{\text{discounting factor asset inv cost}}_{a,y}$ | $\mathbb{R}_{+}$         | $a \in \mathcal{A}$, $y \in \mathcal{Y}$                                                           | Discounting factor for investment cost of a unit of asset $a$ at year $y$                                              | [-]                       |
@@ -149,7 +148,7 @@ In addition, the following subsets represent methods for incorporating additiona
 
 | Name                                    | Domain           | Domains of Indices                                             | Description                                                       | Units           |
 | --------------------------------------- | ---------------- | -------------------------------------------------------------- | ----------------------------------------------------------------- | --------------- |
-| $p^{\text{efficiency}}_{a,y}$           | $\mathbb{R}_{+}$ | $a \in \mathcal{A^{\text{cv}}}$, $y \in \mathcal{Y}$           | efficiency of conversion asset $a$ at year $y$                    | [p.u.]          |
[…]
