### 2025-11-20 f231185
Apply suggestions from code review
Co-authored-by: Lauren Clisby <lclisby@gmail.com>
