### 2025-11-20 457c99b
added standard deviation for times
