### 2025-11-21 cb8512f
Add sugestions from review: create CTE
