### 2025-11-21 cbf3b86
Add sugestions from review: improve profile agg
