### 2025-11-21 fd550b4
Add sugestions from review: improve objective agg query
