### 2025-11-24 3b09e1e
[2S-SO]: Two-stage stochastic optimization (#1417)
* Update case study files

* Update duration calculation to account for the scenario column

* Update the preparation of the profiles structure with scenario index

* Update calculation of expression over clustered year with scenario index

* Update the profile structure to have the scenario index

* Update storage variable over clustered year creation

* Update constraints over clustered year creation

* Update objective function weight calculation

* Update stochastic examples tests

* Update relevant MPS files

* Add sugestions from review:
[…]
