+++ src/objective.jl
@@ -209,12 +209,14 @@ function add_objective!(connection, model, variables, expressions, model_paramet
 
     indices = DuckDB.query(
         connection,
-        "WITH rp_weight AS (
+        "WITH rp_weight_prob AS (
             SELECT
                 year,
                 rep_period,
-                SUM(weight) AS weight_sum
-            FROM rep_periods_mapping
+                SUM(rpm.weight * scn.probability) AS total_weight_prob
+            FROM rep_periods_mapping AS rpm
+            LEFT JOIN stochastic_scenario AS scn
+                ON rpm.scenario = scn.scenario
             GROUP BY year, rep_period
         ),
         rp_res AS (
@@ -228,19 +230,19 @@ function add_objective!(connection, model, variables, expressions, model_paramet
         SELECT
             var.id,
             obj.weight_for_operation_discounts
-                * rp_weight.weight_sum
+                * rp_weight_prob.total_weight_prob
                 * rp_res.resolution
                 * (var.time_block_end - var.time_block_start + 1)
                 * obj.total_variable_cost
-                AS cost,
+                AS cost
         FROM var_flow AS var
         LEFT JOIN t_objective_flows as obj
             ON var.from_asset = obj.from_asset
             AND var.to_asset = obj.to_asset
             AND var.year = obj.milestone_year
-        LEFT JOIN rp_weight
-            ON var.year = rp_weight.year
-            AND var.rep_period = rp_weight.rep_period
+        LEFT JOIN rp_weight_prob
+            ON var.year = rp_weight_prob.year
+            AND var.rep_period = rp_weight_prob.rep_period
         LEFT JOIN rp_res
             ON var.year = rp_res.year
             AND var.rep_period = rp_res.rep_period
@@ -259,12 +261,14 @@ function add_objective!(connection, model, variables, expressions, model_paramet
 
     indices = DuckDB.query(
         connection,
-        "WITH rp_weight AS (
+        "WITH rp_weight_prob AS (
             SELECT
                 year,
                 rep_period,
-                SUM(weight) AS weight_sum
-            FROM rep_periods_mapping
+                SUM(rpm.weight * scn.probability) AS total_weight_prob
+            FROM rep_periods_mapping AS rpm
+            LEFT JOIN stochastic_scenario AS scn
+                ON rpm.scenario = scn.scenario
             GROUP BY year, rep_period
         ),
         rp_res AS (
@@ -289,7 +293,7 @@ function add_objective!(connection, model, variables, expressions, model_paramet
         SELECT
             var.id,
             vint_obj.weight_for_operation_discounts
-                * rp_weight.weight_sum
+                * rp_weight_prob.total_weight_prob
                 * rp_res.resolution
                 * (var.time_block_end - var.time_block_start + 1)
                 * vint_obj.total_variable_cost AS cost
@@ -299,9 +303,9 @@ function add_objective!(connection, model, variables, expressions, model_paramet
             AND var.to_asset = vint_obj.to_asset
             AND var.milestone_year = vint_obj.milestone_year
             AND var.commission_year = vint_obj.commission_year
-        LEFT JOIN rp_weight
-            ON var.milestone_year = rp_weight.year
-            AND var.rep_period = rp_weight.rep_period
+        LEFT JOIN rp_weight_prob
+            ON var.milestone_year = rp_weight_prob.year
+            AND var.rep_period = rp_weight_prob.rep_period
         LEFT JOIN rp_res
             ON var.milestone_year = rp_res.year
             AND var.rep_period = rp_res.rep_period
@@ -319,12 +323,14 @@ function add_objective!(connection, model, variables, expressions, model_paramet
 
     indices = DuckDB.query(
         connection,
[…]
