### 2025-11-25 15cd07d
Update parameters in the tutorials
