### 2025-11-25 74a5bbe
Update efficiencies in the rest of the repository
