### 2025-11-25 9c6d6fe
Apply MD060 rule to align markdown tables (#1437)
* Initial plan

* Apply MD060 rule: align markdown tables in formulation.md and paper.md

Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>
