+++ docs/src/40-scientific-foundation/40-formulation.md
@@ -22,23 +22,23 @@ Depth = [2,3]
 
 In addition, the following asset sets represent methods for incorporating additional variables and constraints in the model.
 
-| Name                                      | Description                                                   | Elements | Superset                                                                                         | Notes                                                                                                                                                                                                                                                           |
-| ----------------------------------------- | ------------------------------------------------------------- | -------- | ------------------------------------------------------------------------------------------------ | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| $\mathcal{A}^{\text{i}}_y$                | Energy assets with investment method at year $y$              |          | $\mathcal{A}^{\text{i}}_y  \subseteq \mathcal{A}$                                                |                                                                                                                                                                                                                                                                 |
-| $\mathcal{A}^{\text{operation}}$          | Energy assets with operation mode at year $y$                 |          | $\mathcal{A}^{\text{operation}}  \subseteq \mathcal{A}$                                               |                                                                                                                                                                                                                                                                 |
-| $\mathcal{A}^{\text{simple investment}}$  | Energy assets with simple investment method at year $y$       |          | $\mathcal{A}^{\text{simple investment}}  \subseteq \mathcal{A}$                                  |                                                                                                                                                                                                                                                                 |
-| $\mathcal{A}^{\text{compact investment}}$ | Energy assets with compact investment method at year $y$      |          | $\mathcal{A}^{\text{compact investment}}  \subseteq \mathcal{A}$                                 | |
-| $\mathcal{A}^{\text{semi-compact investment}}$  | Energy assets with semi-compact investment method at year $y$       |          | $\mathcal{A}^{\text{semi-compact investment}}  \subseteq \mathcal{A}$                                  | |
-| $\mathcal{A}^{\text{ss}}_y$               | Energy assets with seasonal method at year $y$                |          | $\mathcal{A}^{\text{ss}}_y \subseteq \mathcal{A}$                                                | This set contains assets that use the seasonal method method. Please visit the how-to sections for [seasonal storage](@ref seasonal-setup) and [maximum/minimum outgoing energy limit](@ref max-min-outgoing-energy-setup) to learn how to set up this feature. |
-| $\mathcal{A}^{\text{se}}_y$               | Storage energy assets with energy method at year $y$          |          | $\mathcal{A}^{\text{se}}_y \subseteq \mathcal{A}^{\text{s}}$                                     | This set contains storage assets that use investment energy method. Please visit the [how-to section](@ref storage-investment-setup) to learn how to set up this feature.                                                                                       |
-| $\mathcal{A}^{\text{sb}}_y$               | Storage energy assets with binary method at year $y$          |          | $\mathcal{A}^{\text{sb}}_y \subseteq \mathcal{A}^{\text{s}} \setminus \mathcal{A}^{\text{ss}}_y$ | This set contains storage assets that use an extra binary variable to avoid charging and discharging simultaneously. Please visit the [how-to section](@ref storage-binary-method-setup) to learn how to set up this feature.                                   |
-| $\mathcal{A}^{\text{max e}}_y$            | Energy assets with maximum outgoing energy method at year $y$ |          | $\mathcal{A}^{\text{max e}}_y \subseteq \mathcal{A}$                                             | This set contains assets that use the maximum outgoing energy method. Please visit the [how-to section](@ref max-min-outgoing-energy-setup) to learn how to set up this feature.                                                                                |
[…]
