### 2025-11-25 883bcbb
Add docs section on the rule and how to use the python file
