### 2025-11-25 a89fb40
Fix lint for formulation.md
