+++ docs/src/40-scientific-foundation/40-formulation.md
@@ -146,9 +146,9 @@ In addition, the following subsets represent methods for incorporating additiona
 
 #### Extra Parameters for Conversion Assets
 
-| Name                                    | Domain           | Domains of Indices                                             | Description                                                       | Units           |
-| --------------------------------------- | ---------------- | -------------------------------------------------------------- | ----------------------------------------------------------------- | --------------- |
-| $p^{\text{conversion efficiency}}_{a,y}$           | $\mathbb{R}_{+}$ | $a \in \mathcal{A^{\text{cv}}}$, $y \in \mathcal{Y}$           | efficiency of conversion asset $a$ at year $y$                    | [p.u.]          |
+| Name                                     | Domain           | Domains of Indices                                             | Description                                                       | Units           |
+| ---------------------------------------- | ---------------- | -------------------------------------------------------------- | ----------------------------------------------------------------- | --------------- |
+| $p^{\text{conversion efficiency}}_{a,y}$ | $\mathbb{R}_{+}$ | $a \in \mathcal{A^{\text{cv}}}$, $y \in \mathcal{Y}$           | efficiency of conversion asset $a$ at year $y$                    | [p.u.]          |
 
 #### Extra Parameters for Energy Constraints
 
@@ -170,27 +170,27 @@ In addition, the following subsets represent methods for incorporating additiona
 
 ### Parameters for Flows
 
-| Name                                                | Domain           | Domains of Indices                                                                                 | Description                                                                                                           | Units          |
-| --------------------------------------------------- | ---------------- | -------------------------------------------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------- | -------------- |
-| $p^{\text{operational cost}}_{f,y}$                 | $\mathbb{R}_{+}$ | $f \in \mathcal{F}$, $y \in \mathcal{Y}$                                                           | Operational cost of flow $f$ at year $y$                                                                              | [kEUR/MWh]     |
-| $p^{\text{reactance}}_{f,y}$                        | $\mathbb{R}_{+}$ | $f \in \mathcal{F}$, $y \in \mathcal{Y}$                                                           | Reactance of flow $f$ at year $y$                                                                                     | [p.u.]         |
-| $p^{\text{commodity price}}_{f,y}$                          | $\mathbb{R}_{+}$         | $f \in \mathcal{F}$, $y \in \mathcal{Y}$                                                           | Price of the commodity to produce a unit of energy for flow $f$ at year $y$. Only applies to producers.                                                                      | [kEUR/MWh] or other units      |
-| $p^{\text{capacity coefficient}}_{f,y}$             | $\mathbb{R}_{+}$ | $f \in \mathcal{F}$, $y \in \mathcal{Y}$                                                           | Coefficient that multiplies the flow $f$ at year $y$ in the capacity constraints                                      | [-]            |
-| $p^{\text{conversion coefficient}}_{f,y}$           | $\mathbb{R}_{+}$ | $f \in \mathcal{F}$, $y \in \mathcal{Y}$                                                           | Coefficient that multiplies the flow $f$ at year $y$ in the conversion constraints                                    | [-]            |
[…]
