+++ utils/scripts/change-a-column-header.jl
@@ -3,25 +3,33 @@ using DuckDB
 
 root_dir = joinpath(@__DIR__, "..", "..")
 test_inputs = joinpath(root_dir, "test", "inputs")
-dirs = readdir(test_inputs; join = true)
+tutorial_inputs =
+    joinpath(root_dir, "tutorials/inputs/docs/src/10-tutorials/my-awesome-energy-system")
+
+test_dirs = readdir(test_inputs; join = true)
+tutorial_dirs = readdir(tutorial_inputs; join = true)
+dirs = vcat(test_dirs, tutorial_dirs)
 push!(dirs, joinpath(root_dir, "benchmark", "EU"))
 
 for dir in dirs
     if !isdir(dir)
         continue
     end
-    println("Processing directory: $dir")
 
     # Explicit name of the table
-    filename = joinpath(dir, "flow-milestone.csv")
+    filename = joinpath(dir, "asset-commission.csv")
 
     connection = DBInterface.connect(DuckDB.DB)
     _q(s) = DuckDB.query(connection, s)
 
     _q("CREATE TABLE t AS FROM read_csv('$filename')")
 
-    # Add a new column with a default value
-    _q("ALTER TABLE t ADD COLUMN commodity_price DOUBLE DEFAULT 0")
+    # Change the column name
+    try
+        _q("ALTER TABLE t RENAME COLUMN efficiency TO conversion_efficiency")
+    catch err
+        @info "Skipping rename; column not found" filename exception = err
+    end
 
     _q("COPY t TO '$filename' (HEADER, DELIMITER ',')")
 end
