+++ utils/scripts/delete-a-column.jl
@@ -3,7 +3,12 @@ using DuckDB
 
 root_dir = joinpath(@__DIR__, "..", "..")
 test_inputs = joinpath(root_dir, "test", "inputs")
-dirs = readdir(test_inputs; join = true)
+tutorial_inputs =
+    joinpath(root_dir, "tutorials/inputs/docs/src/10-tutorials/my-awesome-energy-system")
+
+test_dirs = readdir(test_inputs; join = true)
+tutorial_dirs = readdir(tutorial_inputs; join = true)
+dirs = vcat(test_dirs, tutorial_dirs)
 push!(dirs, joinpath(root_dir, "benchmark", "EU"))
 
 for dir in dirs
@@ -13,7 +18,7 @@ for dir in dirs
     println("Processing directory: $dir")
 
     # Explicit name of the table
-    filename = joinpath(dir, "profiles-rep-periods.csv")
+    filename = joinpath(dir, "asset-milestone.csv")
 
     connection = DBInterface.connect(DuckDB.DB)
     _q(s) = DuckDB.query(connection, s)
@@ -21,7 +26,11 @@ for dir in dirs
     _q("CREATE TABLE t AS FROM read_csv('$filename')")
 
     # Delete a column by name
-    _q("ALTER TABLE t DROP COLUMN scenario")
+    try
+        _q("ALTER TABLE t DROP COLUMN commodity_price")
+    catch err
+        @info "Skipping drop; column not found" filename exception = err
+    end
 
     _q("COPY t TO '$filename' (HEADER, DELIMITER ',')")
 end
