+++ .github/workflows/CompareMPS.yml
@@ -43,7 +43,7 @@ jobs:
             Please wait.
           edit-mode: replace
       - name: Clone
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
       - name: Setup Julia
         uses: julia-actions/setup-julia@v2
         with:
