+++ .github/workflows/Benchmark.yml
@@ -41,7 +41,7 @@ jobs:
 
             Benchmark in progress...
           edit-mode: replace
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v6
       - uses: julia-actions/setup-julia@latest
         with:
           version: 1
