### 2025-11-26 23e6e7e
Update docs/src/10-tutorials/31-rolling-horizon.md
Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>
