### 2025-11-26 c9aeaaa
Update docs/src/10-tutorials/31-rolling-horizon.md
Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>
