### 2025-11-27 f2c29ff
Update docs/src/10-tutorials/31-rolling-horizon.md
Co-authored-by: Lauren Clisby <lclisby@gmail.com>
