### 2025-12-01 3842286
[WIP] Plot error
