### 2025-12-01 e338235
[WIP] Print dual errors
