### 2025-12-01 7bc0c9d
Update docs/src/10-tutorials/31-rolling-horizon.md
Co-authored-by: Diego Alejandro Tejada Arango <12887482+datejada@users.noreply.github.com>
