### 2025-12-01 50f38d4
Rolling horizon tutorial (#1387)
* Create tutorial for rolling horizon

Closes #1385
Part of #1365

* Update comparison section to show both rolling and non-rolling solutions

* Apply suggestions from code review

Co-authored-by: Lauren Clisby <lclisby@gmail.com>

* Update docs/src/10-tutorials/31-rolling-horizon.md

Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>

* Update docs/src/10-tutorials/31-rolling-horizon.md

Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>

* Update docs/src/10-tutorials/31-rolling-horizon.md

Co-authored-by: Abel Soares S
[…]
