+++ docs/src/10-tutorials/31-rolling-horizon.md
@@ -1,0 +1,489 @@
+# Tutorial 8: Rolling Horizon
+
+In this example we will replicate the [rolling horizon example in JuMP](https://jump.dev/JuMP.jl/stable/tutorials/algorithms/rolling_horizon/).
+Rolling horizon is a technique to try to manage very large problems by trading some accuracy for efficiency.
+Instead of solving the model for the full time horizon model, it solves many smaller problems of the same size, in what are called "rolling windows".
+Each window corresponds to a shorter time period, so it's faster to solve, but since we cannot look at the full horizon, the solution is (potentially) sub-optimal.
+In particular, long-term storage can be affected by the limited information.
+
+For this example we will use the [Rolling Horizon](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/blob/main/test/inputs/Rolling%20Horizon/) test files.
+If you are following along, make sure to download all of these files beforehand.
+
+```@example rolling_horizon
+using DuckDB: DBInterface, DuckDB
+using TulipaIO: TulipaIO
+using TulipaEnergyModel: TulipaEnergyModel
+
+# Define rolling_horizon_folder as the path to where you download the data
+rolling_horizon_folder = joinpath(@__DIR__, "..", "..", "..", "test", "inputs", "Rolling Horizon") # hide
+connection = DBInterface.connect(DuckDB.DB)
+schemas = TulipaEnergyModel.schema_per_table_name
+TulipaIO.read_csv_folder(connection, rolling_horizon_folder; schemas)
+```
+
+To better visualize the data, we will create a helper function `nice_query`.
+This is optional.
+
+```@example rolling_horizon
+using DataFrames
+
+nice_query(sql) = DuckDB.query(connection, sql) |> DataFrame
+```
+
+In this data, we have four assets: `solar`, `thermal`, `battery` and `demand`.
+You can see the connections between them in the `flow` table:
+
+```@example rolling_horizon
+nice_query("SELECT from_asset, to_asset FROM flow")
+```
+
+## Solution without rolling horizon
+
+Let's first solve this problem directly and inspect the solution
+
+```@example rolling_horizon
+energy_problem = TulipaEnergyModel.run_scenario(connection, show_log=false)
+```
+
+We will aggregate the flow solution and visualize the "solar" and "thermal" assets, and the "charge" and "discharge" of the battery.
+
+```@example rolling_horizon
+using Plots: Plots
+
+big_table_no_rh = nice_query("""
+    WITH cte_outgoing AS (
+        SELECT
+            var.from_asset AS asset,
+            var.time_block_start AS timestep,
+            SUM(var.solution) AS solution,
+        FROM var_flow AS var
+        WHERE rep_period = 1 AND year = 2030
+        GROUP BY asset, timestep
+    ), cte_incoming AS (
+        SELECT
+            var.to_asset AS asset,
+            var.time_block_start AS timestep,
+            SUM(var.solution) AS solution,
+        FROM var_flow AS var
+        WHERE rep_period = 1 AND year = 2030
+        GROUP BY asset, timestep
+    ), cte_unified AS (
+        SELECT
+            cte_outgoing.asset,
+            cte_outgoing.timestep,
+            coalesce(cte_outgoing.solution, 0.0) AS outgoing,
+            coalesce(cte_incoming.solution, 0.0) AS incoming,
+        FROM cte_outgoing
+        LEFT JOIN cte_incoming
+            ON cte_outgoing.asset = cte_incoming.asset
+            AND cte_outgoing.timestep = cte_incoming.timestep
+    ) FROM cte_unified
+""")
+
[…]
