### 2025-12-02 a267f70
Release 0.19.0
