### 2025-12-08 e04daa1
[WIP] Working on unit commitment for consumers
