### 2025-12-10 7c7a516
Update days before adding the stale label
