+++ .github/workflows/Stale.yml
@@ -9,13 +9,13 @@ jobs:
   stale:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/stale@v10.1.0
+      - uses: actions/stale@v10.1.1
         with:
           stale-issue-message: This issue is stale because it has been open 2 months with no activity. Remove the stale label or comment, otherwise this will be closed in 5 days.
           stale-pr-message: This PR is stale because it has been open 1 month with no activity. Remove stale label or comment to revive it.
           stale-issue-label: stale
           exempt-issue-labels: "bug"
-          days-before-issue-stale: 60
-          days-before-pr-stale: 30
+          days-before-issue-stale: 180
+          days-before-pr-stale: 180
           days-before-close: 7
           days-before-pr-close: 7
