### 2025-12-10 a0e5438
case study data profiles and timeframe
