+++ debugging/debug.jl
@@ -31,3 +31,5 @@ conn = DBInterface.connect(DuckDB.DB)
 read_csv_folder(conn, in_dir; schemas = TulipaEnergyModel.schema_per_table_name)
 
 energy_problem = run_scenario(conn; model_file_name = "model.lp", log_file = "log_file.log")
+
+# FILL_IN,producer,,FILL_IN,FILL_IN,simple,true,15,15,0.05,==,0.0,false,,true,FILL_IN,true,true,false,0.0,false,FILL_IN,FILL_IN,FILL_IN,FILL_IN,FILL_IN,FILL_IN,FILL_IN,FILL_IN,FILL_IN,FILL_IN
