### 2025-12-10 e3d034c
[WIP] Undo min_output change and order profiles
