### 2025-12-15 46c0359
[WIP] Working on tests for bidding
