+++ test/test-bids.jl
@@ -1,0 +1,232 @@
+@testsnippet BidsSetup begin
+    function create_bids_tables!(connection, bids)
+        # Bids
+        # bidding_window = 1:24
+        lookup = Dict(
+            (bid.customer, bid.exclusive_group, bid.profile_block) => bid_id for
+            (bid_id, bid) in enumerate(bids)
+        )
+        bids_df = DataFrame([
+            (;
+                bid_id = lookup[bid.customer, bid.exclusive_group, bid.profile_block],
+                asset = "bid$(lookup[bid.customer, bid.exclusive_group, bid.profile_block])",
+                bid.customer,
+                bid.exclusive_group,
+                bid.profile_block,
+                bid.price,
+            ) for bid in bids
+        ])
+        bids_profiles_df = DataFrame([
+            (;
+                bid_id = lookup[bid.customer, bid.exclusive_group, bid.profile_block],
+                profile_name = "bid_profiles-bid$(lookup[bid.customer, bid.exclusive_group, bid.profile_block])-demand",
+                timestep = ti,
+                quantity = qi,
+            ) for bid in bids for (ti, qi) in zip(bid.timestep, bid.quantity)
+        ])
+        DuckDB.register_table(connection, bids_df, "bids")
+        DuckDB.register_table(connection, bids_profiles_df, "bids_profiles")
+
+        return nothing
+    end
+end
+
+@testitem "Making bids" setup = [CommonSetup, BidsSetup] tags = [:case_study, :integration, :slow] begin
+    # Normal input
+    dir = joinpath(INPUT_FOLDER, "Rolling Horizon")
+    connection = DBInterface.connect(DuckDB.DB)
+    _read_csv_folder(connection, dir)
+
+    sql(str) = DuckDB.query(connection, str)
+    insert_into(table, what) = sql("INSERT INTO $table BY NAME ($what)")
+    insert_into(table, columns, from) = insert_into(table, "SELECT $columns FROM $from")
+    from_bids_insert_into(table, columns) = insert_into(table, columns, "bids")
+
+    # Normal solution
+    energy_problem = TulipaEnergyModel.run_scenario(connection; show_log = false)
+    @test energy_problem.objective_value ≈ 410873.9 rtol = 1e-5
+
+    # New connection
+    close(connection)
+    connection = DBInterface.connect(DuckDB.DB)
+    _read_csv_folder(connection, dir)
+
+    bid_blocks = [
+        (
+            customer = "A",
+            exclusive_group = 1,
+            profile_block = 1,
+            timestep = 9:9,
+            quantity = [10],
+            price = 500.0,
+        ),
+        (
+            customer = "A",
+            exclusive_group = 2,
+            profile_block = 1,
+            timestep = 7:8,
+            quantity = [70, 30],
+            price = 1.5,
+        ),
+        (
+            customer = "A",
+            exclusive_group = 2,
+            profile_block = 2,
+            timestep = 7:8,
+            quantity = [80, 20],
+            price = 1.5,
+        ),
+        (
+            customer = "B",
+            exclusive_group = 3,
+            profile_block = 1,
+            timestep = 7:13,
+            quantity = (1:7) .* (7:-1:1),
+            price = 0.001,
+        ),
+    ]
+    create_bids_tables!(connection, bid_blocks)
+
+    # Modifications to make bids work
+    year =
+        [
+            row.year for row in
+            DuckDB.query(connection, "SELECT DISTINCT milestone_year AS year FROM asset_milestone")
+        ] |> only
+    timestep_window = only([
+        row for row in DuckDB.query(
+            connection,
[…]
