+++ .github/workflows/Lint.yml
@@ -39,7 +39,7 @@ jobs:
       - name: Hack for setup-python cache # https://github.com/actions/setup-python/issues/807
         run: rm requirements.txt
       - name: Cache pre-commit
-        uses: actions/cache@v4
+        uses: actions/cache@v5
         with:
           path: ~/.cache/pre-commit
           key: ${{ runner.os }}-pre-commit-${{ hashFiles('**/.pre-commit-config.yaml') }}
