### 2026-01-03 e4c805d
Create new test for two assets sharing inflows profile
Test that two assets can share the same inflows_profile in the assets_profiles
table, but have different values of storage_inflows in the asset_milestone
table. Verify that the correct right hand side is created in the storage
balance constraints.

Relates to #1423.
Related to an intermitent issue in the OBZ tutorial
