### 2026-01-03 9f936ff
Don't include asset_milestone.storage_inflows in the inflows profile
Since the inflows profile are only tied to the profile_name, we have to delay
multiplying by storage_inflows until the assets (i.e., the constraints), are
involved.
