### 2026-01-03 3b7b992
Further cleanup of creating inflows profiles
