### 2026-01-05 6a17e37
Add compat bounds to docs and test Project.toml
