### 2026-01-05 5a4bfab
New test and fix for inflows profiles (#1439)
* Create new test for two assets sharing inflows profile

Test that two assets can share the same inflows_profile in the assets_profiles
table, but have different values of storage_inflows in the asset_milestone
table. Verify that the correct right hand side is created in the storage
balance constraints.

Relates to #1423.
Related to an intermitent issue in the OBZ tutorial

* Don't include asset_milestone.storage_inflows in the inflows profile

Since the inflows profile are only tied to the profile_name, we have to delay
multiplying by storage_inf
[…]
