### 2026-01-05 48ab988
Germán's update reducing around 200 words (#1448)
Co-authored-by Germán Morales-España <german.morales@tno.nl>
