### 2026-01-05 bc2cafd
Replace CompatHelper by dependabot for compat autoupdate
Closes #1449
