+++ benchmark/Project.toml
@@ -4,3 +4,12 @@ DuckDB = "d2f5444f-75bc-4fdf-ac35-56f514c445e1"
 MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
 TulipaEnergyModel = "5d7bd171-d18e-45a5-9111-f1f11ac5d04d"
 TulipaIO = "7b3808b7-0819-42d4-885c-978ba173db11"
+
+[sources]
+TulipaEnergyModel = {path = ".."}
+
+[compat]
+BenchmarkTools = "1"
+DuckDB = "1"
+MetaGraphsNext = "0.7"
+TulipaIO = "0.5"
