+++ Project.toml
@@ -44,4 +44,4 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 test = ["Test"]
 
 [workspaces]
-projects = ["test", "docs"]
+projects = ["test", "docs", "benchmark"]
