### 2026-01-05 38596bb
Replace CompatHelper by dependabot for compat autoupdate (#1450)
Closes #1449
