+++ .github/dependabot.yml
@@ -5,3 +5,11 @@ updates:
     directory: "/" # Location of package manifests
     schedule:
       interval: "weekly"
+  - package-ecosystem: "julia"
+    directories:
+      - "/"
+      - "/benchmark"
+      - "/docs"
+      - "/test"
+    schedule:
+      interval: "weekly"
