### 2026-01-05 c2b3eea
Update TulipaEnergyModel requirement to 0.19.0 in /benchmark
Updates the requirements on [TulipaEnergyModel](https://github.com/TulipaEnergy/TulipaEnergyModel.jl) to permit the latest version.
- [Release notes](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/releases)
- [Commits](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/compare/v0.1.0...v0.19.0)

---
updated-dependencies:
- dependency-name: TulipaEnergyModel
  dependency-version: 0.19.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
