+++ benchmark/Project.toml
@@ -12,4 +12,5 @@ TulipaEnergyModel = {path = ".."}
 BenchmarkTools = "1"
 DuckDB = "1"
 MetaGraphsNext = "0.7"
+TulipaEnergyModel = "0.19.0"
 TulipaIO = "0.5"
