+++ docs/Project.toml
@@ -32,4 +32,5 @@ LiveServer = "1"
 OrderedCollections = "1"
 Plots = "1"
 TulipaClustering = "0.4 - 0.5"
+TulipaEnergyModel = "0.19.0"
 TulipaIO = "0.5.0"
