+++ test/Project.toml
@@ -33,4 +33,5 @@ TestItemRunner = "1"
 TestItems = "1"
 TulipaBuilder = "0.1"
 TulipaClustering = "0.5"
+TulipaEnergyModel = "0.19.0"
 TulipaIO = "0.5"
