### 2026-01-05 94fba5b
Update TOML requirement to 1.0.3 in /test
Updates the requirements on [TOML](https://github.com/JuliaLang/TOML.jl) to permit the latest version.
- [Release notes](https://github.com/JuliaLang/TOML.jl/releases)
- [Commits](https://github.com/JuliaLang/TOML.jl/commits)

---
updated-dependencies:
- dependency-name: TOML
  dependency-version: 1.0.3
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
