### 2026-01-06 aee3705
Ignore TulipaEnergyModel version in dependabot (#1455)
