+++ test/Project.toml
@@ -29,7 +29,7 @@ GLPK = "1"
 HiGHS = "1"
 JuMP = "1"
 MathOptInterface = "1"
-TOML = "1.0.3"
+TOML = "1"
 TestItemRunner = "1"
 TestItems = "1"
 TulipaBuilder = "0.1"
