### 2026-01-06 2f5498f
Update TOML requirement to 1 in /test (#1454)
* Update TOML requirement to 1.0.3 in /test

Updates the requirements on [TOML](https://github.com/JuliaLang/TOML.jl) to permit the latest version.
- [Release notes](https://github.com/JuliaLang/TOML.jl/releases)
- [Commits](https://github.com/JuliaLang/TOML.jl/commits)

---
updated-dependencies:
- dependency-name: TOML
  dependency-version: 1.0.3
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>

* Update test/Project.toml

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: de
[…]
