### 2026-01-12 #9447 @kyleniemeyer
Thanks @marlinarnz! And just FYI, commands to `editorialbot` need to come at the beginning of a comment for it to see them.
