### 2026-01-19 2b324dd
Update docs on pre-commit (#1459)
