+++ src/expressions/multi-year.jl
@@ -304,7 +304,7 @@ function _create_multi_year_expressions_indices!(connection, expressions)
         WHERE
             asset.investment_method in ('simple', 'none')
             -- Hub and consumer assets do not use this expression, so we can filter them out to be more explicit
-            AND asset.type in ('producer', 'conversion', 'storage')
+            AND asset.type in ('producer', 'conversion', 'storage', 'consumer')
         GROUP BY asset_both.asset, asset_both.milestone_year, asset_both.commission_year
         ",
     )
