### 2026-01-19 71b88ec
Test that bids work as expected
Test the previous commit. The testing has two parts.
First, we have a setup step that performs all the
necessary modifications in the input data to create the bids
as Tulipa-compliant input.
Second, we have the sets of tests. The first set checks
that the generated constraints are correct.
The second set checks that the solution is correct for many
cases of capacity and operational_cost.
The third set checks that a case with curtailment is correct.
