+++ test/test-bids.jl
@@ -1,0 +1,593 @@
+@testsnippet BidsSetup begin
+    using TulipaBuilder
+    using TulipaClustering: TulipaClustering
+
+    insert_into(connection, table, what) =
+        DuckDB.query(connection, "INSERT INTO $table BY NAME ($what)")
+    insert_into(connection, table, columns, from) =
+        insert_into(connection, table, "SELECT $columns FROM $from")
+    from_bids_insert_into(connection, table, columns) =
+        insert_into(connection, table, columns, "bids")
+
+    bid_blocks = [
+        (
+            customer = "A",
+            exclusive_group = 1,
+            profile_block = 1,
+            timestep = 6:6,
+            quantity = [10],                # 0, 0, 10, ....
+            price = 500.0,
+            curtailment_minimum = 1.0,
+        ),
+        (
+            customer = "A",
+            exclusive_group = 2,
+            profile_block = 1,
+            timestep = 4:5,
+            quantity = [70, 30],            # 70, 30, 0, ....
+            price = 2.5,
+            curtailment_minimum = 1.0,
+        ),
+        (
+            customer = "A",
+            exclusive_group = 2,
+            profile_block = 2,
+            timestep = 4:5,
+            quantity = [60, 20],            # 60, 20, 0, ...
+            price = 1.5,
+            curtailment_minimum = 0.8,
+        ),
+        (
+            customer = "B",
+            exclusive_group = 1,
+            profile_block = 1,
+            timestep = 4:10,
+            quantity = (1:7) .* (7:-1:1),   # 7, 12, 15, 16, 15, 12, 7
+            price = 0.001,
+            curtailment_minimum = 1.0,
+        ),
+    ]
+
+    """
+        connection = create_problem(; capacity, operational_cost)
+
+    Creates a simple problem via TulipaBuilder with given capacity and operational_cost.
+
+    The problem has one `producer` named "Generator", one `consumer` named "Bid Manager" and flow between them.
+    The producer has capacity given by the input and `initial_units = 1`.
+    The consumer has `peak_demand = 0`.
+    Since we don't have a profile, we explicitly define a milestone year with 12 timesteps.
+    After this problem is created, we "dummy cluster" it and populate with defaults.
+    """
+    function create_problem(; capacity = 1.0, operational_cost = 1.0)
+        tulipa = TulipaData{String}()
+        add_asset!(tulipa, "Generator", :producer; capacity = capacity, initial_units = 1.0)
+        add_asset!(tulipa, "Bid Manager", :consumer; peak_demand = 0.0) # no demand
+        add_flow!(tulipa, "Generator", "Bid Manager"; operational_cost = operational_cost)
+        # Because we at least one profile
+        attach_profile!(tulipa, "Bid Manager", :demand, 2030, zeros(12))
+
+        connection = create_connection(tulipa)
+        TulipaClustering.dummy_cluster!(connection)
+        TulipaEnergyModel.populate_with_defaults!(connection)
+
+        return connection
+    end
+
+    """
+        create_bids_tables!(connection, bids)
+
+    Given a NamedTuple array `bids` with a format similar to what is created in
+    DEMOSES, we create the `bids` and `bids_profiles` tables in the DuckDB `connection`.
+
+    Each bid has a (consumer, exclusive_group, profile_block) identifier.
+    Under the same `(consumer, exclusive_group)` group, at most one of the `profile_block`s is selected.
+
+    Each bid has two vector of the same size, `timestep` and `quantity`,
+    indicating the bid block's desired quantities per timestep.
+    The whole block sells for the price `price`.
+
+    These tables' schema are:
+    - `bids`: `bid_id INT, asset TEXT, customer TEXT, exclusive_group INT, profile_block INT, price REAL`
[…]
