+++ Project.toml
@@ -1,7 +1,18 @@
 name = "TulipaEnergyModel"
 uuid = "5d7bd171-d18e-45a5-9111-f1f11ac5d04d"
 version = "0.19.1"
-authors = ["Abel Soares Siqueira <abel.s.siqueira@gmail.com>", "Diego A. Tejada-Arango <diego.tejadaarango@tno.nl>", "Germán Morales-España <german.morales@tno.nl>", "Grigory Neustroev <G.Neustroev@tudelft.nl>", "Juha Kiviluoma <Juha.Kiviluoma@vtt.fi>", "Lauren Clisby <lauren.clisby@tno.nl>", "Maaike Elgersma <m.b.elgersma@tudelft.nl>", "Ni Wang <ni.wang@tno.nl>", "Suvayu Ali <s.ali@esciencecenter.nl>", "Zhi Gao <z.gao1@uu.nl>"]
+authors = [
+  "Abel Soares Siqueira <abel.s.siqueira@gmail.com>",
+  "Diego A. Tejada-Arango <diego.tejadaarango@tno.nl>",
+  "Germán Morales-España <german.morales@tno.nl>",
+  "Grigory Neustroev <G.Neustroev@tudelft.nl>",
+  "Juha Kiviluoma <Juha.Kiviluoma@vtt.fi>",
+  "Lauren Clisby <lauren.clisby@tno.nl>",
+  "Maaike Elgersma <m.b.elgersma@tudelft.nl>",
+  "Ni Wang <ni.wang@tno.nl>",
+  "Suvayu Ali <s.ali@esciencecenter.nl>",
+  "Zhi Gao <z.gao1@uu.nl>",
+]
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
