### 2026-01-19 e9571d7
Add validation for bids
Adds a validation check for bids. This defines three sufficient cases
for bids:
- At least 1 consumer asset has unit commitment (since unit commitment
  for consumers cannot be used for anything else at the moment)
- At least 1 loop flow exists
- At least 1 flow with negative operational cost exists
In any of these cases, it is expected that the data contains the
workarounds to create bids. The validation checks this, and we create
tests for these cases.
