+++ src/data-validation.jl
@@ -73,6 +73,11 @@ function validate_data!(connection)
             _validate_flow_commission_and_asset_both_consistency!,
             false,
         ),
+        (
+            "consumer unit commitment used as bids has right data",
+            _validate_bid_related_data!,
+            false,
+        ),
     )
         @timeit to "$log_msg" append!(error_messages, validation_function(connection))
         if fail_fast && length(error_messages) > 0
@@ -592,6 +597,257 @@ function _validate_flow_commission_and_asset_both_consistency!(connection)
             error_messages,
             "Unexpected commission_year = $(row.commission_year) for the outgoing flow of asset '$(row.from_asset)' in 'flow_commission'. The commission_year should match the one in 'asset_both'.",
         )
+    end
+
+    return error_messages
+end
+
+function _validate_bid_related_data!(connection)
+    error_messages = String[]
+
+    #= Testing strategy:
+    # - For a given `asset`, there are necessary and sufficient conditions that
+    #   imply that this asset represents a bid.
+    # - We loop over each sufficient condition and get all assets that satisfy that condition
+    # - For each of these assets, we verify the necessary conditions.
+    =#
+
+    """
+        get_bid_data(connection)
+
+    Gets all relevant data related to an asset in dictionary format. Used to
+    verify the necessary conditions for a bid to be defined correctly:
+
+    - asset.type = 'consumer'
+    - asset.unit_commitment = true
+    - asset.unit_commitment_integer = true
+    - asset.unit_commitment_method = 'basic'
+    - asset.consumer_balance_sense = '=='
+    - asset.capacity = 1.0
+    - asset_both.initial_units = 1.0
+    - assets_rep_periods_partitions.specification = 'uniform' and partition = num_timesteps of rep_period
+    - assets_profiles.type = 'demand' (i.e., there is a 'demand' profile for this asset)
+    - there is a loop flow.from_asset = flow.to_asset = asset.asset (i.e., the asset has a loop) (by itself is already prohibitive)
+    - there is a flow from some consumer to this asset, with operational_cost < 0
+    - there is a single rep_period
+    - there is a single year
+    """
+    function get_bid_data(connection)
+        has_demand_profile_str = "false"
+
+        if _check_if_table_exists(connection, "assets_profiles")
+            has_demand_profile_str = """
+            EXISTS (
+                FROM assets_profiles
+                WHERE assets_profiles.asset = asset.asset
+                    AND assets_profiles.profile_type = 'demand'
+            )"""
+        end
+
+        has_wrong_asset_partition_str = "true" # partition needs to be defined because default is hourly
+        if _check_if_table_exists(connection, "assets_rep_periods_partitions")
+            has_wrong_asset_partition_str = """
+            EXISTS (
+                FROM assets_rep_periods_partitions AS partitions
+                LEFT JOIN rep_periods_data AS rpdata
+                    ON partitions.year = rpdata.year AND partitions.rep_period = rpdata.rep_period
+                WHERE partitions.asset = asset.asset
+                    AND (
+                        partitions.specification != 'uniform'
+                        OR partitions.partition != rpdata.num_timesteps
+                    )
+            )"""
+        end
+
+        query = """
+        SELECT
+            asset.asset,
+            asset.type,
+            asset.unit_commitment,
+            asset.unit_commitment_integer,
+            asset.unit_commitment_method,
+            asset.consumer_balance_sense,
+            asset.capacity,
+            asset_both.initial_units,
+            EXISTS (
+                FROM flow
+                WHERE flow.from_asset = asset.asset
+                    AND flow.to_asset = asset.asset
+            ) AS has_loop,
+            flow_milestone.from_asset as bid_manager,
[…]
