+++ test/test-data-validation.jl
@@ -694,3 +694,179 @@ end
         connection,
     )
 end
+
+@testitem "Check that consumer unit commitment implies other bid-related data to be correct" setup =
+    [CommonSetup] tags = [:unit, :data_validation, :fast] begin
+    using TulipaBuilder
+    using TulipaClustering
+
+    year = 2030
+    year_length = 12
+
+    function create_problem_base()
+        tulipa = TulipaData{String}()
+        add_asset!(tulipa, "Generator", :producer; capacity = 50.0, initial_units = 1.0)
+        add_asset!(tulipa, "Bid Manager", :consumer; peak_demand = 0.0) # no demand
+        add_flow!(tulipa, "Generator", "Bid Manager"; operational_cost = 5.0)
+        # Because we at least one profile
+        attach_profile!(tulipa, "Bid Manager", :demand, 2030, zeros(year_length))
+
+        return tulipa
+    end
+
+    function create_connection_and_prepare(tulipa)
+        connection = create_connection(tulipa)
+        TulipaClustering.dummy_cluster!(connection)
+        TulipaEnergyModel.populate_with_defaults!(connection)
+        return connection
+    end
+
+    """
+        add_new_bid!(tulipa_data, bid_name; price, quantity, ...)
+
+    Creates a new bid using TulipaBuilder for the problem created by `create_problem_base`.
+    The `price` and `quantity` are expected keywords. The extra keywords are used to give
+    incorrect values to the problem. Using these extra keywords, we can create
+    a bid, except for one wrong thing.
+    """
+    function add_new_bid!(
+        tulipa,
+        bid_name;
+        price,
+        quantity,
+        add_demand_profile = true,
+        add_flow_from_bid_manager = true,
+        add_loop = true,
+        capacity = 1.0,
+        consumer_balance_sense = "==",
+        flow_from_bid_manager = true,
+        initial_units = 1.0,
+        set_partition = true,
+        type = :consumer,
+        unit_commitment = true,
+        unit_commitment_integer = true,
+        unit_commitment_method = "basic",
+    )
+        profile = zeros(year_length)
+        for (t, qty) in quantity
+            profile[t] = qty
+        end
+        add_asset!(
+            tulipa,
+            bid_name,
+            type;
+            capacity,
+            consumer_balance_sense,
+            initial_units,
+            min_operating_point = 1.0,
+            peak_demand = maximum(profile),
+            unit_commitment,
+            unit_commitment_integer,
+            unit_commitment_method,
+        )
+        if set_partition
+            set_partition!(tulipa, bid_name, year, 1, year_length)
+        end
+        if add_loop
+            add_flow!(tulipa, bid_name, bid_name)
+        end
+        if add_flow_from_bid_manager
+            add_flow!(tulipa, "Bid Manager", bid_name; operational_cost = -price)
+        end
+        if add_demand_profile
+            attach_profile!(tulipa, bid_name, :demand, year, profile)
+        end
+
+        return tulipa
+    end
+
+    @testset "All data is correct" begin
+        tulipa = create_problem_base()
+        add_new_bid!(tulipa, "bid1"; price = 1.0, quantity = Dict(3 => 30.0, 4 => 50.0))
+        connection = create_connection_and_prepare(tulipa)
+
+        error_messages = TEM._validate_bid_related_data!(connection)
+        @test length(error_messages) == 0
+    end
+
+    """
+        error_messages_for_wrong_problem(; kwargs...)
+
+    Helper function to create a new problem with a bid. The `kwargs` are
+    expected wrong values. Returns the error messages generated by
+    `_validate_bid_related_data`.
+    """
+    function error_messages_for_wrong_problem(; num_years = 1, num_rep_periods = 1, kwargs...)
+        tulipa = create_problem_base()
[…]
