### 2026-01-19 e9a54c4
Create a new case study based on DEMOSES bids
