### 2026-01-19 236f584
Create docs for bids
