+++ docs/src/10-tutorials/40-bids-workaround.md
@@ -1,0 +1,626 @@
+# [Bids via workaround using consumer unit commitment](@id bids-tutorial)
+
+In this tutorial we will learn how to create bids using consumer unit commitment.
+Bids are not explicitly supported in Tulipa (yet), but they can be modeled with a few workarounds.
+
+This is an advanced tutorial. It assumes some basic knowledge of Tulipa, so going through some of the earlier tutorials might be beneficial.
+Furthermore, it deals with the underlying model of Tulipa, so it
+
+## Introduction
+
+In our context, a bid is a proposal to buy energy at a given price at one or more time steps.
+If the proposal is for a single time step, then we are going to call it a "simple bid", as opposed to a "profile bid", when it involves more than one time step.
+The price is constant in both cases, which allow us to always use vectors to represent the required quantities.
+In a "simple bid", the vectors of time steps and quantities both have 1 element.
+
+Furthermore, the bid can be part of an "exclusive group". Inside each exclusive group, a single bid is accepted.
+Finally, some bids also have a curtailment possibility, i.e., they can be supplied with less energy than the maximum desired, but no less than a given percentage of the maximum (given by `curtailment_minimum`).
+
+Here are some example bids:
+
+```@example bids
+bid_blocks = [
+    (
+        customer = "A",
+        exclusive_group = 1,
+        profile_block = 1,
+        timestep = 4:4,
+        quantity = [10],
+        price = 5.0,
+        curtailment_minimum = 1.0,
+    ),
+    (
+        customer = "A",
+        exclusive_group = 2,
+        profile_block = 1,
+        timestep = 2:3,
+        quantity = [40, 30],
+        price = 2.5,
+        curtailment_minimum = 1.0,
+    ),
+    (
+        customer = "A",
+        exclusive_group = 2,
+        profile_block = 2,
+        timestep = 2:3,
+        quantity = [20, 20],
+        price = 1.5,
+        curtailment_minimum = 0.8,
+    ),
+    (
+        customer = "B",
+        exclusive_group = 1,
+        profile_block = 1,
+        timestep = 1:6,
+        quantity = [5, 10, 15, 25, 30, 15],
+        price = 0.8,
+        curtailment_minimum = 1.0,
+    ),
+]
+```
+
+Each bid has the following data:
+
+- `customer`, identifying who is the asking party;
+- `exclusive_group`, identifying each group of exclusive bids;
+- `profile_block`, identifying each block of bids;
+- `timestep`, indicating the time steps of a bid;
+- `quantity`, indicating the vector of requested quantities;
+- `price`, indicating the price;
+- `curtailment_minimum`, indicating the minimum percentage of energy that can be delivered under curtailment.
+
+Notice that `(customer, exclusive_group, profile_block)` form a unique identifier for this bid.
+
+In words, we can say:
+
+- The first bid, `(A, 1, 1)`, requests 10 KW at time step 4 and is willing to pay \$5.0 per KW. No curtailment allowed.
+- The second bid, `(A, 2, 1)`, requests 40 KW at time step 2 and 30 KW at time step 3 and is willing to pay \$2.5 per KW. No curtailment allowed.
+- The third bid, `(A, 2, 2)`, requests 20 KW at time step 2 and 20 KW at time step 3 and is willing to pay \$1.5 per KW. At least 80% of the requested quantity per day must be satisfied.
+- The fourth bid, `(B, 1, 1)`, requests 5 KW, 10 KW, 15 KW, 25 KW, 30 KW, and 15 KW from time steps 1 to 6, in order. It is willing to pay \$0.8 per KW, and no curtailment is allowed.
+
[…]
