### 2026-01-19 b2f7fac
Tutorial Stochastic commit
