### 2026-01-20 03d54f7
Add validation and tests for commodity_price profile
