+++ src/data-validation.jl
@@ -78,6 +78,11 @@ function validate_data!(connection)
             _validate_bid_related_data!,
             false,
         ),
+        (
+            "commodity_price profiles require commodity_price > 0",
+            _validate_commodity_price_consistency!,
+            false,
+        ),
     )
         @timeit to "$log_msg" append!(error_messages, validation_function(connection))
         if fail_fast && length(error_messages) > 0
@@ -854,6 +859,46 @@ function _validate_bid_related_data!(connection)
                     "$prefix_msg has wrong asset partition. It should be uniform and equal to num_timesteps for all representative periods",
                 )
             end
+        end
+    end
+
+    return error_messages
+end
+
+function _validate_commodity_price_consistency!(connection)
+    error_messages = String[]
+    if !_check_if_table_exists(connection, "flows_profiles")
+        return error_messages
+    end
+    # All flows associated with a 'commodity_price' profile
+    rows = [
+        row for row in DuckDB.query(
+            connection,
+            """
+            SELECT
+                flow_milestone.from_asset,
+                flow_milestone.to_asset,
+                flows_profiles.profile_type,
+                flow_milestone.commodity_price,
+            FROM flows_profiles
+            LEFT JOIN flow_milestone
+                ON flow_milestone.from_asset = flows_profiles.from_asset
+                AND flow_milestone.to_asset = flows_profiles.to_asset
+            WHERE
+                flows_profiles.profile_type = 'commodity_price'
+            """,
+        )
+    ]
+    if length(rows) == 0
+        return error_messages # No commodity_price profile, just leave
+    end
+    for row in rows
+        if row.commodity_price <= 0
+            from, to, commodity_price = row.from_asset, row.to_asset, row.commodity_price
+            push!(
+                error_messages,
+                "Flow ($from, $to) is associated with a 'commodity_price' profile, so it should have flow_milestone.commodity_price > 0, but we found $commodity_price",
+            )
         end
     end
 
