### 2026-01-20 f2c09ac
Fix Project.toml
[workspace], not [workspaces]
