+++ Project.toml
@@ -37,11 +37,5 @@ TimerOutputs = "0.5"
 TulipaIO = "0.5"
 julia = "1.10"
 
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
-
-[workspaces]
+[workspace]
 projects = ["test", "docs", "benchmark"]
