### 2026-01-20 cc3e341
Remove unnecessary test and fix test_logs
