### 2026-01-20 332d55c
Add missing test packages and fix test_logs min_level
