### 2026-01-20 0ce1c1d
Remove logging, it doesn't play well with DuckDB warning
