+++ test/test-data-validation.jl
@@ -873,8 +873,6 @@ end
 
 @testitem "Check that commodity_price profile requires flow_milestone.commodity_price > 0" tags =
     [:fast, :validation] setup = [CommonSetup] begin
-    using Logging
-
     connection = _tiny_fixture()
 
     # Add a commodity_price profile and link to (demand, flow)
@@ -909,5 +907,5 @@ end
         WHERE from_asset = 'wind' AND to_asset = 'demand'
         """,
     )
-    @test_logs min_level = Logging.Info TEM.create_internal_tables!(connection)
+    TEM.create_internal_tables!(connection)
 end
