### 2026-01-20 26b3d6e
Try to improve performance by skipping aggregation function when there is no profile
