### 2026-01-20 b51984a
Typing to try to improve performance
