### 2026-01-21 65aa8f1
Fix link in 40-bids-workaround.md
