### 2026-01-21 21f47ca
Testing different strategy in consumer constraints to improve performance
