### 2026-01-21 4d673ae
Add early exit in data validation
