### 2026-01-21 eee8796
Try new strategy for creating profiles structure
