### 2026-01-21 cb48955
Create benchmark helper files and document it
Create tulipa-data.jl with a problem generator via TulipaBuilder.
Create timer-output.jl, benchmarktools.jl, and profview.jl,
showcasing three cases of performance investigation.
Remove old benchmark/profiling.jl file.
Update the docs section for performance.
