### 2026-01-22 #9878 @editorialbot
[REVIEW]: TulipaEnergyModel.jl: A Modelling Framework Breaking the Tradeoff Between Fidelity and Computational Load
**Submitting author:** @abelsiqueira (<a href="http://orcid.org/0000-0003-4451-281X">Abel Soares Siqueira</a>)
**Repository:** https://github.com/TulipaEnergy/TulipaEnergyModel.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.18.2
