### 2026-01-23 4696de5
Test commodity_price validation when there is no flows_profiles table
