### 2026-01-23 f47e252
Improve validation of commodity_price profiles by filtering only violations
