### 2026-01-23 1cca976
Only join flows_profiles in flows_operational_cost if profile exist
In the creation of the indices for flows_operational_cost, only join the
flows_profiles table if the commodity_price profile exists.
