### 2026-01-23 2443cad
Workaround allowing bids via unit commitment for consumers (#1440)
* Modify Tulipa to allow bids to be created via unit commitment for consumers

Modifies Tulipa to create unit commitment for consumers.
Create special conditions when there is a loop flow to connect the loop flow to the unit commitment variables.
Also create min and max output flow with unit commitment variables for consumers.
All these changes make it possible to create bids by:
- Creating a consumer asset with unit commitment
- Create a loop flow in the asset
- Create a flow from an existing consumer asset to the bid asset wi
[…]
