+++ docs/Project.toml
@@ -11,12 +11,13 @@ JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+TulipaBuilder = "05fe4dc7-9a58-53a4-b912-b5673d74e816"
 TulipaClustering = "314fac8b-c762-4aa3-9d12-851379729163"
 TulipaEnergyModel = "5d7bd171-d18e-45a5-9111-f1f11ac5d04d"
 TulipaIO = "7b3808b7-0819-42d4-885c-978ba173db11"
 
 [sources]
-TulipaEnergyModel = {path = ".."}
+TulipaEnergyModel = { path = ".." }
 
 [compat]
 CSV = "0.10"
@@ -31,5 +32,6 @@ JuMP = "1"
 LiveServer = "1"
 OrderedCollections = "1"
 Plots = "1"
+TulipaBuilder = "0.1.2"
 TulipaClustering = "0.4 - 0.5"
 TulipaIO = "0.5.0"
