+++ test/Project.toml
@@ -32,6 +32,6 @@ MathOptInterface = "1"
 TOML = "1"
 TestItemRunner = "1"
 TestItems = "1"
-TulipaBuilder = "0.1"
+TulipaBuilder = "0.1.2"
 TulipaClustering = "0.5"
 TulipaIO = "0.5"

[…]
