### 2026-01-23 e3f160d
Add a few more type hints for commodity_price profile
