### 2026-01-23 9219aa5
Refactor implementation
