### 2026-01-23 d52eee6
Correct flows_operational_cost expression and broaden test
