### 2026-01-23 3e9291a
Revert "Refactor implementation"
This reverts commit 9219aa58b42b9064002b6e29a55dfb9f06a20e86.
