### 2026-01-23 b8b1dea
Fix bug when investment cost is zero (#1465)
* Handle investment_cost=0

* Refactor implementation

* Revert "Refactor implementation"

This reverts commit 9219aa58b42b9064002b6e29a55dfb9f06a20e86.
