### 2026-01-23 8a2d3ca
Allow commodity_price to be a profile (#1410)
* Allow commodity_price to be a profile

Delay computation of cost using commodity_price out of
t_objective_flows.

Part of #1269

* Update src/objective.jl

* Add validation and tests for commodity_price profile

* Remove unnecessary test and fix test_logs

* Add missing test packages and fix test_logs min_level

* Remove logging, it doesn't play well with DuckDB warning

* Try to improve performance by skipping aggregation function when there is no profile

* Typing to try to improve performance

* Test commodity_price validation when there is no
[…]
