### 2026-01-26 6deb39d
Update Statistics requirement to 1.11.1 in /test
Updates the requirements on [Statistics](https://github.com/JuliaStats/Statistics.jl) to permit the latest version.
- [Release notes](https://github.com/JuliaStats/Statistics.jl/releases)
- [Commits](https://github.com/JuliaStats/Statistics.jl/compare/v1.4.0...v1.11.1)

---
updated-dependencies:
- dependency-name: Statistics
  dependency-version: 1.11.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
