### 2026-01-26 #9447 @jamesglynn
> My responses will be delayed until 2026-02-09. In the meantime, [@lbonaldo](https://github.com/lbonaldo), [@jamesglynn](https://github.com/jamesglynn), [@trulsf](https://github.com/trulsf), could you kindly consider your availability for this review? A decline will also help.

I can do a review starting the second week in February. would that be time enough?
