+++ benchmark/scalability.jl
@@ -1,0 +1,170 @@
+using BenchmarkTools: @benchmark
+using CSV: CSV
+using DataFrames: DataFrame
+using DuckDB: DuckDB, DBInterface
+using Plots: Plots
+using Statistics: Statistics
+using TulipaEnergyModel: TulipaEnergyModel as TEM
+using TulipaIO: TulipaIO as TIO
+
+include("tulipa-data.jl")
+
+function common_setup(; kwargs...)
+    connection, tulipa_data = give_me_better_name(; kwargs...)
+    TEM.populate_with_defaults!(connection)
+    return connection, tulipa_data
+end
+
+# using StatsPlots
+
+function setup_lower_level_pipeline(; kwargs...)
+    connection, tulipa_data = common_setup(; kwargs...)
+
+    # Internal data and structures pre-model
+    TEM.create_internal_tables!(connection)
+    model_parameters = TEM.ModelParameters(connection)
+    variables = TEM.compute_variables_indices(connection)
+    constraints = TEM.compute_constraints_indices(connection)
+    profiles = TEM.prepare_profiles_structure(connection)
+
+    #= Comment out before revelant function
+    # Create model
+    model, expressions = TEM.create_model(
+        connection,
+        variables,
+        constraints,
+        profiles,
+        model_parameters,
+    )
+
+    # Solve model
+    TEM.solve_model(model)
+    TEM.save_solution!(connection, model, variables, constraints)
+    output_dir = mktempdir()
+    TEM.export_solution_to_csv_files(output_dir, connection)
+    =#
+
+    return connection
+end
+
+function relevant_lower_level_pipeline(connection)
+    # Write relevant function
+    model, expressions =
+        TEM.create_model(connection, variables, constraints, profiles, model_parameters)
+
+    return nothing
+end
+
+function setup_higher_level_pipeline(; kwargs...)
+    connection, tulipa_data = common_setup(; kwargs...)
+
+    energy_problem = TEM.EnergyProblem(connection)
+    #= Comment out before revelant function
+    TEM.create_model!(energy_problem)
+    TEM.solve_model!(energy_problem)
+    TEM.save_solution!(energy_problem)
+    TEM.export_solution_to_csv_files(mktempdir(), energy_problem)
+    =#
+
+    return connection, energy_problem
+end
+
+function relevant_higher_level_pipeline(connection, energy_problem)
+    # Write relevant function
+    TEM.create_model!(energy_problem)
+
+    return nothing
+end
+
+function compute_results(grid)
+    results = Dict()
+    for num_countries in grid.num_countries, num_rep_periods in grid.num_rep_periods
+        @info "Running benchmark for num_countries = $num_countries and num_rep_periods = $num_rep_periods"
+        problem_kwargs = (; num_rep_periods, num_countries)
+        key = (num_countries, num_rep_periods)
+
+        # Uncomment one of the two
+        # Lower level API
+        # results[key] = @benchmark relevant_lower_level_pipeline(connection) setup=(connection=setup_lower_level_pipeline(;$problem_kwargs...))
+
+        # Higher level API
+        results[key] = @benchmark relevant_higher_level_pipeline(connection, energy_problem) setup =
+            ((connection, energy_problem) = setup_higher_level_pipeline(; $problem_kwargs...))
+
+        @info results[key]
+    end
+
+    return results
+end
+
+function fake_results(grid)
+    results = Dict()
+    for num_countries in grid.num_countries, num_rep_periods in grid.num_rep_periods
[…]
