### 2026-01-30 37cdff8
Update CompareMPS to not comment on PR when run by dependabot (#1468)
Dependabot doesn't have permissions to create pull request comments, so
we skip these steps in CompareMPS when the actor is dependabot.
Verify that mps-logfile.txt exists, just in case something else creates
the same kind of issue.

Related to the failure in #1466
