### 2026-01-30 9142fba
[Fix]: Update Statistics dependency version to 1 (#1470)
