+++ test/Project.toml
@@ -30,7 +30,7 @@ GLPK = "1"
 HiGHS = "1"
 JuMP = "1"
 MathOptInterface = "1"
-Statistics = "1.11.1"
+Statistics = "1"
 TOML = "1"
 TestItemRunner = "1"
 TestItems = "1"
