### 2026-01-30 cb4ec09
Add memory-profile.jl script to help investigate memory usage
Related to #457
