### 2026-01-30 ac15dae
Add memory-profile.jl script to help investigate memory usage
Related to #457
