### 2026-01-30 c2ef96f
Try to improve memory usage in add_capacity_constraints
