### 2026-02-02 59a7920
Small type stability fixes resulting from JET investigation
