+++ src/input-schemas.jl
@@ -1,6 +1,7 @@
 # read schema from file
 
-const schema = JSON.parsefile(joinpath(@__DIR__, "input-schemas.json"); dicttype = OrderedDict);
+const schema =
+    JSON.parsefile(joinpath(@__DIR__, "input-schemas.json"); dicttype = OrderedDict{String,Any});
 
 const schema_per_table_name = OrderedDict(
     schema_key => OrderedDict(key => value["type"] for (key, value) in schema_content) for
