### 2026-02-02 19e0b69
Add type-stability.jl to help investigate type instability issues
Example of how to use JET to investigate type instability
issues, and developer documentation about it.

Related to #459
