### 2026-02-02 8dd39e6
Move input table schemas outside workflow Close #1289
