+++ docs/src/10-tutorials/9-stochastic.md
@@ -1,28 +1,344 @@
 # Tutorial 9: Stochastic method
 
-## Introduction 
+## Introduction
 
-Stochastic programming methods are used to model the uncertainty in long-term investment decisions explicitly by considering multiple scenarios. In a two-stage stochastic setting, the investment decision is the first stage decision and the operational decisions after uncertainty is realized represent the second stage.
+Stochastic programming methods are used to model the uncertainty in long-term investment decisions explicitly by considering multiple scenarios. In a two-stage stochastic setting, the investment decision is modeled as the first stage decision and the operational decisions after uncertainty is realized represent the second stage.
 
 The TulipaClustering.jl package allows the creation of blended Representative Periods (RPs) to reduce the size of the problem. A tutorial on this package is available under [Tutorial4](https://tulipaenergy.github.io/TulipaEnergyModel.jl/stable/10-tutorials/15-clustering-rep-periods/)
 
 In the stochastic setting, RPs can be clustered Per or Cross scenario. Here is the concept documentation for more detail: [Clustering per or Cross](https://tulipaenergy.github.io/TulipaClustering.jl/stable/20-concepts/#Clustering-Per-or-Cross)
 
-## Previously in the TLC 
+## Previously in the TLC
 
 We reuse the instantiation from previous tutorials, and subsequently use the data of tutorial 9
-
 
 ```julia
 using Pkg: Pkg
 Pkg.activate(".")
-Pkg.instantiate() # only if update packages
+Pkg.instantiate()
+Pkg.add("TulipaClustering")
+Pkg.add("Distances") # only if update packages
 import TulipaIO as TIO
 import TulipaEnergyModel as TEM #forces TEM. before calling functions
+import TulipaClustering as TC
+import CSV
 using DuckDB
 using DataFrames
+import Distances
 using Plots
+using Statistics
 
+connection = DBInterface.connect(DuckDB.DB)
+input_dir = "docs/src/10-tutorials/my-awesome-energy-system/tutorial-9"
+output_dir = "docs/src/10-tutorials/my-awesome-energy-system/tutorial-9/results"
+TIO.read_csv_folder(connection, input_dir)
 ```
 
+## Profiles for Stochastic
 
+Show the profiles `profiles-wide`.
+
+```julia
+profiles = TIO.get_table(connection, "profiles_wide")
+```
+
+Observe a DataFrame of 8760 time steps per scenario. Group the DataFrame by year and scenario and summarize the capacities by taking the mean.
+
+```julia
+gdf = groupby(profiles, [:year, :scenario])
+result_df = combine(gdf, [:solar, :wind_offshore, :wind_onshore, :demand, :hydro_inflow] .=> mean)
+```
+
+There are 3 different weather years and you see the difference in the average capacities of the availability of the RESs and demand in different scenarios.
+
+## Clustering Per Scenario
+
+Transform the data from wide to long and
+
+```julia
+TC.transform_wide_to_long!(
+    connection,
+    "profiles_wide",
+    "profiles";
+    exclude_columns=["scenario", "year", "timestep"],
+)
+```
+
+> **Note:** Make the stochastic methods comparable on level of detail by ensuring that the num_rps is not divisible by the number of scenarios.
+
+```julia
+n_scenarios = length(unique(profiles_wide.scenario))
+layout = TC.ProfilesTableLayout(; cols_to_groupby=[:year, :scenario])
+num_rps = 8
+period_duration = 24
+clusters = TC.cluster!(connection,
+                    period_duration,
+                    num_rps;
+                    method = :convex_hull,
+                    distance = Distances.CosineDist(),
[…]
