### 2026-02-02 af0c013
Change energy_capacity to capacity_storage_energy
