### 2026-02-02 f8a2ce7
Remove numbers from tutorials
