### 2026-02-02 a351727
Move input table schemas outside workflow (#1474)
Close #1289
