+++ docs/src/20-user-guide/54-input-table-schemas.md
@@ -1,0 +1,66 @@
+# [Input Table Schemas](@id table-schemas)
+
+The input data must follow the table schemas below to correctly build a system in Tulipa.
+
+The schemas below are in [`input-schemas.json`](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/blob/main/src/input-schemas.json). You can also view the schemas after loading the package by typing `TulipaEnergyModel.schema` in the Julia console.
+
+!!! info "Optional tables/files and their defaults"
+    The following tables/files are allowed to be missing: "assets\_rep\_periods\_partitions", "assets\_timeframe\_partitions", "assets\_timeframe\_profiles", "flows\_rep\_periods\_partitions", "group\_asset", "profiles\_timeframe".
+    - For the partitions tables/files, the default value are `specification = uniform` and `partition = 1` for each asset/flow and year
+    - For the profiles tables/files, the default value is a flat profile of value 1.0 p.u.
+    - If no group table/file is available there will be no group constraints in the model
+
+```@eval
+"""
+The output of the following code is a Markdown text with the following structure:
+
+TABLE_NAME
+=========
+
+PARAMETER_NAME
+
+  •  Description: Lorem ipsum
+  •  Type: SQL type of the parameter
+  •  Default: a value or "No default"
+  •  Unit of measure: a value or "No unit"
+  •  Constraints: a table or "No constraints"
+"""
+
+using Markdown, JSON
+using OrderedCollections: OrderedDict
+
+input_schemas = JSON.parsefile("../../../src/input-schemas.json"; dicttype = OrderedDict)
+
+let buffer = IOBuffer()
+    for (i,(table_name, fields)) in enumerate(input_schemas)
+        write(buffer, "## Table $i : `$table_name`\n\n")
+        for (field_name, field_info) in fields
+            _description = get(field_info, "description", "No description provided")
+            _type = get(field_info, "type", "Unknown type")
+            _unit = get(field_info, "unit_of_measure", "No unit")
+            _default = get(field_info, "default", "No default")
+            _constraints_values = get(field_info, "constraints", nothing)
+
+            write(buffer, "**`$field_name`**\n\n")
+            write(buffer, "- Description: $_description\n\n")
+            write(buffer, "- Type: `$_type`\n")
+            write(buffer, "- Unit of measure: `$_unit` \n")
+            write(buffer, "- Default: `$_default`\n")
+
+            if _constraints_values === nothing
+                write(buffer, "- Constraints: No constraints\n")
+            elseif isa(_constraints_values, OrderedDict)
+                write(buffer, "| Constraints | Value |\n| --- | --- |\n")
+                for (key, value) in _constraints_values
+                    write(buffer, "| $key | `$value` |\n")
+                end
+                write(buffer, "\n")
+            else
+                write(buffer, "- Constraints: `$(string(_constraints_values))`\n")
+            end
+        end
+    end
+    Markdown.parse(String(take!(buffer)))
+end
+
+```
