### 2026-02-02 fd165ae
Move FFTR contraints table out of Concepts Close #1378
