### 2026-02-02 5225902
Add test for storage energy balance
