+++ test/test-constraint-storage-balance.jl
@@ -1,0 +1,157 @@
+@testitem "Test storage balance constraints - single scenario" setup = [CommonSetup] tags =
+    [:unit, :validation, :fast] begin
+    using DuckDB
+    using JuMP
+    using TulipaEnergyModel: TulipaEnergyModel as TEM
+    using TulipaClustering: TulipaClustering as TC
+    using TulipaBuilder: TulipaBuilder as TB
+
+    # We create a simple tulipa problem with two storage assets
+    tulipa = TB.TulipaData()
+
+    storage_data = Dict(
+        "storageA" => Dict(
+            "inflows" => 10,
+            "is_seasonal" => true,
+            "initial_units" => 1,
+            "initial_storage_units" => 1,
+            "storage_charging_efficiency" => 0.85,
+            "storage_discharging_efficiency" => 0.8,
+            "initial_storage_level" => 0.5,
+            "capacity" => 1,
+            "capacity_storage_energy" => 168,
+        ),
+        "storageB" => Dict(
+            "inflows" => 0.0,
+            "is_seasonal" => false,
+            "initial_units" => 1,
+            "initial_storage_units" => 1,
+            "storage_charging_efficiency" => 0.9,
+            "storage_discharging_efficiency" => 0.95,
+            "initial_storage_level" => 0.0,
+            "capacity" => 1,
+            "capacity_storage_energy" => 4,
+        ),
+    )
+    TB.add_asset!(tulipa, "generator", :producer)
+    TB.add_asset!(tulipa, "consumer", :consumer)
+    TB.add_flow!(tulipa, "generator", "consumer")
+    TB.add_flow!(tulipa, "consumer", "storageB")
+    TB.add_flow!(tulipa, "storageB", "consumer")
+
+    for storage_asset in ("storageA", "storageB")
+        TB.add_asset!(
+            tulipa,
+            storage_asset,
+            :storage;
+            storage_inflows = storage_data[storage_asset]["inflows"],
+            is_seasonal = storage_data[storage_asset]["is_seasonal"],
+            initial_units = storage_data[storage_asset]["initial_units"],
+            initial_storage_units = storage_data[storage_asset]["initial_storage_units"],
+            storage_charging_efficiency = storage_data[storage_asset]["storage_charging_efficiency"],
+            storage_discharging_efficiency = storage_data[storage_asset]["storage_discharging_efficiency"],
+            initial_storage_level = storage_data[storage_asset]["initial_storage_level"],
+            capacity = storage_data[storage_asset]["capacity"],
+            capacity_storage_energy = storage_data[storage_asset]["capacity_storage_energy"],
+        )
+        TB.add_flow!(tulipa, "consumer", storage_asset)
+        TB.add_flow!(tulipa, storage_asset, "consumer")
+    end
+
+    # Attach profile to storageA
+    inflows_profile = [1; 5.5; 10]
+    periods = 1:length(inflows_profile)
+    TB.attach_profile!(tulipa, "storageA", :inflows, 2030, inflows_profile)
+
+    connection = TB.create_connection(tulipa)
+
+    # Cluster, populate_with_defaults, and create model
+    period_duration = 1
+    num_rps = 2
+    clusters = TC.cluster!(
+        connection,
+        period_duration,
+        num_rps;
+        method = :convex_hull,
+        weight_type = :convex,
+        tol = 1e-6,
+        weight_fitting_kwargs = Dict(:learning_rate => 0.001, :niters => 1000),
+    )
+    TEM.populate_with_defaults!(connection)
+    energy_problem = TEM.EnergyProblem(connection)
+    TEM.create_model!(energy_problem; model_file_name = joinpath(@__DIR__, "test_storage.lp"))
+
+    # Test for the StorageA
[…]
