+++ docs/src/10-tutorials/15-clustering-rep-periods.md
@@ -1,4 +1,4 @@
-# Tutorial 4: Representative Periods with Tulipa Clustering
+# Blended Representative Periods with Tulipa Clustering
 
 ## Introduction
 
