### 2026-02-03 7bc94e4
Rename a storage parameter (#1475)
* Change energy_capacity to capacity_storage_energy

* Minor edit
