### 2026-02-03 65722e7
Update MetaGraphsNext requirement from 0.7 to 0.7, 0.8 in /benchmark (#1479)
Updates the requirements on [MetaGraphsNext](https://github.com/JuliaGraphs/MetaGraphsNext.jl) to permit the latest version.
- [Release notes](https://github.com/JuliaGraphs/MetaGraphsNext.jl/releases)
- [Commits](https://github.com/JuliaGraphs/MetaGraphsNext.jl/compare/v0.7.0...v0.8.0)

---
updated-dependencies:
- dependency-name: MetaGraphsNext
  dependency-version: 0.8.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot
[…]
