### 2026-02-03 5120025
chore: :robot: pre-commit update (#1471)
Co-authored-by: abelsiqueira <1068752+abelsiqueira@users.noreply.github.com>
