### 2026-02-03 fec1536
Fix Julia Formatter pre-commit hook to use julia command directly (#1481)
* Initial plan

* Fix Julia Formatter pre-commit hook to use julia command directly

Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>
