### 2026-02-03 9d349f1
Remove hub asset from docs.
