+++ docs/src/30-concepts.md
@@ -9,13 +9,13 @@ Depth = [2, 3]
 
 _TulipaEnergyModel.jl_ uses two fundamental building blocks as the foundation of the optimization model:
 
-- **Assets**: representation of a physical asset that can produce, consume, store, balance, or convert energy. Some examples of what these assets can represent are:
+- **Assets**: representation of a physical asset that can produce, consume, balance, store, or convert energy. Some examples of what these assets can represent are:
   - Producer: e.g., wind turbine, solar panel
-  - Consumer: e.g., electricity demand, heat demand
+  - Consumer: e.g., electricity demand, heat demand, energy balance
+    - Note that a consumer can have 0 demand and serve as a balance or "hub"
   - Storage: e.g., battery, pumped-hydro storage
-  - Balancing Hub: e.g., an electricity network that serves as a connection among other energy assets
   - Conversion: e.g., power plants, electrolyzers
-- **Flows**: representation of the connections among assets, e.g., pipelines, transmission lines, or simply the energy production that goes from one asset to another.
+- **Flows**: connections between assets, e.g., pipelines, transmission lines, or simply the energy production that goes from one asset to another.
 
 In a nutshell, the model guarantees a balance of energy for the various types of assets while considering the flow limits. It considers a set of [representative periods](@ref representative-periods) (e.g., days or weeks) for a given [timeframe](@ref timeframe) (e.g., a year) the user wants to analyze. Therefore, the model has two types of temporal (time) constraints to consider the different chronology characteristics of the assets:
 
@@ -42,7 +42,7 @@ Consider the following example to demonstrate the benefits of using a graph theo
 
 ![Classic connection](./figs/flexible-connection-1.png)
 
-In this system, the `phs` storage asset charges and discharges from the `connection point`, while the `wind` turbine produces power that goes directly to the `connection point`. This `connection point` is connected to the external power grid through a transmission line that leads to a `balance` hub that connects to other assets. Essentially, the `connection point` acts as a balancing hub point for the assets in this hybrid configuration. Furthermore, these hybrid configurations impose an extra constraint to avoid storage charges from the power grid.
+In this system, the `phs` storage asset charges and discharges from the `connection point`, while the `wind` turbine produces power that goes directly to the `connection point`. This `connection point` is connected to the external power grid through a transmission line that leads to a `balance` consumer (0 demand) that connects to other assets. Essentially, the `connection point` acts as a balancing point for the assets in this hybrid configuration. Furthermore, these hybrid configurations impose an extra constraint to avoid storage charges from the power grid.
 
 Let's consider the modeling approach in _TulipaEnergyModel.jl_. As nodes are no longer needed to connect assets, we can connect them directly to each other, as shown in the figure below:
 
@@ -62,7 +62,7 @@ Let's zoom in on the `phs-wind-balance` triangle and see what happens in the fig
 
 !!! warning "Be careful with the definition of flows"
     1. Although transport flows are bidirectional, they must be defined in a single direction. For example, a producer like wind can only have outgoing flows. Thus, the flow between `wind` and `balance` must be specified as the flow from `wind` to `balance`, with its sign allowed to be free.
[…]
