### 2026-02-03 73749e4
Add test for non seasonal storage
