+++ test/test-constraint-storage-balance.jl
@@ -1,156 +1,424 @@
-@testitem "Test storage balance constraints - single scenario" setup = [CommonSetup] tags =
-    [:unit, :validation, :fast] begin
-    using DuckDB
-    using JuMP
-    using TulipaEnergyModel: TulipaEnergyModel as TEM
-    using TulipaClustering: TulipaClustering as TC
+@testsnippet StorageSetup begin
+    using DuckDB: DuckDB
     using TulipaBuilder: TulipaBuilder as TB
+    using TulipaClustering: TulipaClustering as TC
 
-    # We create a simple tulipa problem with two storage assets
-    tulipa = TB.TulipaData()
-
-    storage_data = Dict(
-        "storageA" => Dict(
-            "inflows" => 10,
+    # Storage asset configuration data
+    const STORAGE_CONFIGS = Dict{String,Dict{String,Union{Float64,Bool,Int}}}(
+        "seasonal_storage" => Dict{String,Union{Float64,Bool,Int}}(
+            "inflows" => 10.0,
             "is_seasonal" => true,
-            "initial_units" => 1,
-            "initial_storage_units" => 1,
+            "initial_units" => 1.0,
+            "initial_storage_units" => 1.0,
             "storage_charging_efficiency" => 0.85,
             "storage_discharging_efficiency" => 0.8,
             "initial_storage_level" => 0.5,
-            "capacity" => 1,
-            "capacity_storage_energy" => 168,
+            "capacity" => 1.0,
+            "capacity_storage_energy" => 168.0,
         ),
-        "storageB" => Dict(
-            "inflows" => 0.0,
+        "non_seasonal_storage" => Dict{String,Union{Float64,Bool,Int}}(
+            "inflows" => 3.5,
             "is_seasonal" => false,
-            "initial_units" => 1,
-            "initial_storage_units" => 1,
+            "initial_units" => 1.0,
+            "initial_storage_units" => 1.0,
             "storage_charging_efficiency" => 0.9,
             "storage_discharging_efficiency" => 0.95,
-            "initial_storage_level" => 0.0,
-            "capacity" => 1,
-            "capacity_storage_energy" => 4,
+            "initial_storage_level" => 0.25,
+            "capacity" => 1.0,
+            "capacity_storage_energy" => 4.0,
         ),
     )
-    TB.add_asset!(tulipa, "generator", :producer)
-    TB.add_asset!(tulipa, "consumer", :consumer)
-    TB.add_flow!(tulipa, "generator", "consumer")
-    TB.add_flow!(tulipa, "consumer", "storageB")
-    TB.add_flow!(tulipa, "storageB", "consumer")
 
-    for storage_asset in ("storageA", "storageB")
-        TB.add_asset!(
-            tulipa,
-            storage_asset,
-            :storage;
-            storage_inflows = storage_data[storage_asset]["inflows"],
-            is_seasonal = storage_data[storage_asset]["is_seasonal"],
-            initial_units = storage_data[storage_asset]["initial_units"],
-            initial_storage_units = storage_data[storage_asset]["initial_storage_units"],
-            storage_charging_efficiency = storage_data[storage_asset]["storage_charging_efficiency"],
-            storage_discharging_efficiency = storage_data[storage_asset]["storage_discharging_efficiency"],
-            initial_storage_level = storage_data[storage_asset]["initial_storage_level"],
-            capacity = storage_data[storage_asset]["capacity"],
-            capacity_storage_energy = storage_data[storage_asset]["capacity_storage_energy"],
+    """
+        create_storage_balance_problem(; inflows_profile=[1.0, 5.5, 10.0], period_duration=1, num_rps=2)
+
+    Create a storage balance test problem with two storage assets (seasonal_storage and non_seasonal_storage).
[…]
