### 2026-02-03 8a34946
Remove hub from OBZ example
