+++ test/test-constraint-storage-balance.jl
@@ -3,37 +3,50 @@
     using TulipaBuilder: TulipaBuilder as TB
     using TulipaClustering: TulipaClustering as TC
 
+    # Type-stable storage configuration struct
+    struct StorageConfig
+        inflows::Float64
+        is_seasonal::Bool
+        initial_units::Float64
+        initial_storage_units::Float64
+        charging_efficiency::Float64
+        discharging_efficiency::Float64
+        initial_storage_level::Float64
+        capacity::Float64
+        capacity_storage_energy::Float64
+    end
+
     # Storage asset configuration data
-    const STORAGE_CONFIGS = Dict{String,Dict{String,Union{Float64,Bool,Int}}}(
-        "seasonal_storage" => Dict{String,Union{Float64,Bool,Int}}(
-            "inflows" => 10.0,
-            "is_seasonal" => true,
-            "initial_units" => 1.0,
-            "initial_storage_units" => 1.0,
-            "storage_charging_efficiency" => 0.85,
-            "storage_discharging_efficiency" => 0.8,
-            "initial_storage_level" => 0.5,
-            "capacity" => 1.0,
-            "capacity_storage_energy" => 168.0,
+    const STORAGE_CONFIGS = Dict{String,StorageConfig}(
+        "seasonal_storage" => StorageConfig(
+            10.0,  # inflows
+            true,  # is_seasonal
+            1.0,   # initial_units
+            1.0,   # initial_storage_units
+            0.85,  # charging_efficiency
+            0.8,   # discharging_efficiency
+            0.5,   # initial_storage_level
+            1.0,   # capacity
+            168.0, # capacity_storage_energy
         ),
-        "non_seasonal_storage" => Dict{String,Union{Float64,Bool,Int}}(
-            "inflows" => 3.5,
-            "is_seasonal" => false,
-            "initial_units" => 1.0,
-            "initial_storage_units" => 1.0,
-            "storage_charging_efficiency" => 0.9,
-            "storage_discharging_efficiency" => 0.95,
-            "initial_storage_level" => 0.25,
-            "capacity" => 1.0,
-            "capacity_storage_energy" => 4.0,
+        "non_seasonal_storage" => StorageConfig(
+            3.5,   # inflows
+            false, # is_seasonal
+            1.0,   # initial_units
+            1.0,   # initial_storage_units
+            0.9,   # charging_efficiency
+            0.95,  # discharging_efficiency
+            0.25,  # initial_storage_level
+            1.0,   # capacity
+            4.0,   # capacity_storage_energy
         ),
     )
 
     """
-        create_storage_balance_problem(; inflows_profile=[1.0, 5.5, 10.0], period_duration=1, num_rps=2)
+        create_storage_balance_problem(; inflows_profile, period_duration, num_rps)
 
-    Create a storage balance test problem with two storage assets (seasonal_storage and non_seasonal_storage).
-    Returns the connection and energy_problem ready for constraint testing.
+    Create a storage balance test problem with two storage assets.
+    Returns the database connection with configured storage assets and clustering.
     """
     function create_storage_balance_problem(;
         inflows_profile::Dict{Tuple{String,Int},Vector{Float64}} = Dict(
@@ -41,33 +54,33 @@
         ),
         period_duration::Int = 1,
         num_rps::Int = 2,
-    )
-        # Create tulipa data structure
+    )::DuckDB.DB
         tulipa = TB.TulipaData()
 
-        # Add basic producer and consumer
+        # Add basic producer and consumer for flow balance
         TB.add_asset!(tulipa, "generator", :producer)
         TB.add_asset!(tulipa, "consumer", :consumer)
         TB.add_flow!(tulipa, "generator", "consumer")
 
[…]
