### 2026-02-06 27def22
Update test with 2 scenarios
