+++ test/test-constraint-storage-balance.jl
@@ -49,8 +49,8 @@
     Returns the database connection with configured storage assets and clustering.
     """
     function create_storage_balance_problem(;
-        inflows_profile::Dict{Tuple{String,Int},Vector{Float64}} = Dict(
-            ("seasonal_storage", 2030) => [1.0, 5.5, 10.0],
+        inflows_profile::Dict{Tuple{String,Int,Int},Vector{Float64}} = Dict(
+            ("seasonal_storage", 2030, 1) => [1.0, 5.5, 10.0],
         ),
         period_duration::Int = 1,
         num_rps::Int = 2,
@@ -81,16 +81,24 @@
             TB.add_flow!(tulipa, "consumer", storage_name)
             TB.add_flow!(tulipa, storage_name, "consumer")
 
-            # Attach inflows profiles for this storage asset
-            for ((asset, year), values) in inflows_profile
+            # Attach inflows profiles per  storage asset, year and scenario
+            for ((asset, year, scenario), values) in inflows_profile
                 if asset == storage_name
-                    TB.attach_profile!(tulipa, storage_name, :inflows, year, values)
+                    TB.attach_profile!(
+                        tulipa,
+                        storage_name,
+                        :inflows,
+                        year,
+                        values;
+                        scenario = scenario,
+                    )
                 end
             end
         end
 
         # Create connection and apply clustering
         connection = TB.create_connection(tulipa)
+        layout = TC.ProfilesTableLayout(; cols_to_crossby = [:scenario])
         TC.cluster!(
             connection,
             period_duration,
@@ -99,6 +107,7 @@
             weight_type = :convex,
             tol = 1e-6,
             weight_fitting_kwargs = Dict(:learning_rate => 0.001, :niters => 1000),
+            layout = layout,
         )
 
         return connection
@@ -144,6 +153,40 @@
     end
 
     """
+        get_storage_ids(connection, storage_asset, scenarios, periods) -> (storage_level_ids)
+
+    Query storage level over clustered year IDs for a storage asset for each scenario and period block.
+    Returns a dictionary indexed by (scenario, period).
+    """
+    function get_storage_ids(
+        connection::DuckDB.DB,
+        storage_asset::String,
+        scenarios::UnitRange{Int},
+        periods::UnitRange{Int},
+    )::Dict{Tuple{Int,Int},Int}
+        # Pre-allocate dictionaries
+        storage_level_ids = Dict{Tuple{Int,Int},Int}(
+            (scenario, period) => 0 for scenario in scenarios for period in periods
+        )
+
+        # Query storage level IDs for each scenario and period block
+        for s in scenarios, p in periods
+            storage_level_ids[(s, p)] = [
+                row.id for row in DuckDB.query(
+                    connection,
+                    "SELECT id
+                     FROM var_storage_level_over_clustered_year
+                     WHERE asset = '$storage_asset' AND
+                           scenario = $s AND
+                           period_block_start = $p",
+                )
+            ][1] # one variable per scenario and period block
+        end
+
+        return storage_level_ids
+    end
+
+    """
         get_inflow_value(profiles_rep_periods, rp, tb) -> Float64
 
     Extract inflow value for a given representative period and time block.
@@ -180,8 +223,9 @@
         flow::Vector{JuMP.VariableRef},
         incoming_flow_ids::Dict{Tuple{Int,Int},Vector{Int}},
         outgoing_flow_ids::Dict{Tuple{Int,Int},Vector{Int}},
[…]
