### 2026-02-06 94c5827
Improve readability, type stability, efficiency
