+++ test/test-constraint-storage-balance.jl
@@ -118,6 +118,8 @@
 
     Query incoming and outgoing flow IDs for a storage asset across all time blocks.
     Returns two dictionaries indexed by (rep_period, time_block_start).
+
+    Optimized to use a single batched query instead of multiple queries.
     """
     function get_flow_ids(
         connection::DuckDB.DB,
@@ -126,37 +128,51 @@
         period_duration::Int,
     )::Tuple{Dict{Tuple{Int,Int},Vector{Int}},Dict{Tuple{Int,Int},Vector{Int}}}
         # Pre-allocate dictionaries for all time blocks
-        incoming = Dict{Tuple{Int,Int},Vector{Int}}(
-            (rp, tb) => Int[] for rp in 1:num_rps for tb in 1:period_duration
-        )
-        outgoing = Dict{Tuple{Int,Int},Vector{Int}}(
-            (rp, tb) => Int[] for rp in 1:num_rps for tb in 1:period_duration
-        )
+        incoming = Dict{Tuple{Int,Int},Vector{Int}}()
+        outgoing = Dict{Tuple{Int,Int},Vector{Int}}()
 
-        # Query flow IDs for each time block
+        # Initialize empty vectors for each time block
         for rp in 1:num_rps, tb in 1:period_duration
-            incoming[(rp, tb)] = [
-                row.id for row in DuckDB.query(
-                    connection,
-                    "SELECT id FROM var_flow WHERE to_asset = '$storage_asset' AND rep_period = $rp AND time_block_start = $tb",
-                )
-            ]
-            outgoing[(rp, tb)] = [
-                row.id for row in DuckDB.query(
-                    connection,
-                    "SELECT id FROM var_flow WHERE from_asset = '$storage_asset' AND rep_period = $rp AND time_block_start = $tb",
-                )
-            ]
+            incoming[(rp, tb)] = Int[]
+            outgoing[(rp, tb)] = Int[]
+        end
+
+        # Batch query for both incoming and outgoing flows
+        query_incoming = """
+            SELECT id, rep_period, time_block_start
+            FROM var_flow
+            WHERE to_asset = '$storage_asset'
+            ORDER BY rep_period, time_block_start
+        """
+        query_outgoing = """
+            SELECT id, rep_period, time_block_start
+            FROM var_flow
+            WHERE from_asset = '$storage_asset'
+            ORDER BY rep_period, time_block_start
+        """
+
+        # Process incoming flows
+        for row in DuckDB.query(connection, query_incoming)
+            key = (Int(row.rep_period), Int(row.time_block_start))
+            push!(incoming[key], row.id)
+        end
+
+        # Process outgoing flows
+        for row in DuckDB.query(connection, query_outgoing)
+            key = (Int(row.rep_period), Int(row.time_block_start))
+            push!(outgoing[key], row.id)
         end
 
         return (incoming, outgoing)
     end
 
     """
-        get_storage_ids(connection, storage_asset, scenarios, periods) -> (storage_level_ids)
+        get_storage_ids(connection, storage_asset, scenarios, periods) -> storage_level_ids
 
     Query storage level over clustered year IDs for a storage asset for each scenario and period block.
     Returns a dictionary indexed by (scenario, period).
+
+    Optimized to use a single batched query instead of multiple queries.
     """
     function get_storage_ids(
         connection::DuckDB.DB,
@@ -164,23 +180,20 @@
         scenarios::UnitRange{Int},
         periods::UnitRange{Int},
     )::Dict{Tuple{Int,Int},Int}
-        # Pre-allocate dictionaries
-        storage_level_ids = Dict{Tuple{Int,Int},Int}(
-            (scenario, period) => 0 for scenario in scenarios for period in periods
-        )
+        storage_level_ids = Dict{Tuple{Int,Int},Int}()
[…]
