### 2026-02-07 c72e613
Move performance scripts to benchmark/profiling and finishing touches
