+++ benchmark/Project.toml
@@ -1,25 +1,18 @@
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DuckDB = "d2f5444f-75bc-4fdf-ac35-56f514c445e1"
-JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
-MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
-PProf = "e4faabce-9ead-11e9-39d9-4379958e3056"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-ProfileView = "c46f51b8-102a-5cf2-8d2c-8597cb0e0da7"
-StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
-TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 TulipaBuilder = "05fe4dc7-9a58-53a4-b912-b5673d74e816"
 TulipaClustering = "314fac8b-c762-4aa3-9d12-851379729163"
 TulipaEnergyModel = "5d7bd171-d18e-45a5-9111-f1f11ac5d04d"
 TulipaIO = "7b3808b7-0819-42d4-885c-978ba173db11"
 
 [sources]
-TulipaEnergyModel = {path = ".."}
+TulipaEnergyModel = { path = ".." }
+
+[workspace]
+projects = ["profiling"]
 
 [compat]
 BenchmarkTools = "1"
 DuckDB = "1"
-MetaGraphsNext = "0.7, 0.8"
 TulipaIO = "0.5"
