### 2026-02-09 77c2ad8
Update stale issue and PR messages for 6 months (#1482)
