+++ .github/workflows/Stale.yml
@@ -11,8 +11,8 @@ jobs:
     steps:
       - uses: actions/stale@v10.1.1
         with:
-          stale-issue-message: This issue is stale because it has been open 2 months with no activity. Remove the stale label or comment, otherwise this will be closed in 5 days.
-          stale-pr-message: This PR is stale because it has been open 1 month with no activity. Remove stale label or comment to revive it.
+          stale-issue-message: This issue is stale because it has been open 6 months with no activity. Remove the stale label or comment to revive it; otherwise this will be closed in 7 days.
+          stale-pr-message: This PR is stale because it has been open 6 month with no activity. Remove the stale label or comment to revive it; otherwise this will be closed in 7 days.
           stale-issue-label: stale
           exempt-issue-labels: "bug"
           days-before-issue-stale: 180
