### 2026-02-10 c5cd0ea
Apply suggestions from code review
Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>
