### 2026-02-10 d4f4ce5
Apply suggestions from code review
Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
