+++ benchmark/profiling/profview.jl
@@ -7,7 +7,7 @@ using TulipaIO: TulipaIO as TIO
 include("../tulipa-data.jl")
 
 function common_setup(; kwargs...)
-    connection, tulipa_data = give_me_better_name(; kwargs...)
+    connection, tulipa_data = create_synthetic_problem(; kwargs...)
     reset_timer!(TEM.to)
     TEM.populate_with_defaults!(connection)
     return connection, tulipa_data
@@ -19,7 +19,7 @@ function setup_lower_level_pipeline(; kwargs...)
     # Internal data and structures pre-model
     TEM.create_internal_tables!(connection)
     model_parameters = TEM.ModelParameters(connection)
-    #= Comment out before revelant function
+    #= Comment out before relevant function
     variables = TEM.compute_variables_indices(connection)
     constraints = TEM.compute_constraints_indices(connection)
     profiles = TEM.prepare_profiles_structure(connection)
@@ -52,7 +52,7 @@ function setup_higher_level_pipeline(; kwargs...)
     connection, tulipa_data = common_setup(; kwargs...)
 
     energy_problem = TEM.EnergyProblem(connection)
-    #= Comment out before revelant function
+    #= Comment out before relevant function
     TEM.create_model!(energy_problem)
     TEM.solve_model!(energy_problem)
     TEM.save_solution!(energy_problem)
