### 2026-02-10 e3ebfa3
Updates to the benchmarks and profiling (#1467)
* Create benchmark helper files and document it

Create tulipa-data.jl with a problem generator via TulipaBuilder.
Create timer-output.jl, benchmarktools.jl, and profview.jl,
showcasing three cases of performance investigation.
Remove old benchmark/profiling.jl file.
Update the docs section for performance.

* Initial approach to testing scalability

* Expand benchmarks to include lower level pipeline

* Add memory-profile.jl script to help investigate memory usage

Related to #457

* Add type-stability.jl to help investigate type instability iss
[…]
