+++ benchmark/Project.toml
@@ -1,15 +1,18 @@
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 DuckDB = "d2f5444f-75bc-4fdf-ac35-56f514c445e1"
-MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
+TulipaBuilder = "05fe4dc7-9a58-53a4-b912-b5673d74e816"
+TulipaClustering = "314fac8b-c762-4aa3-9d12-851379729163"
 TulipaEnergyModel = "5d7bd171-d18e-45a5-9111-f1f11ac5d04d"
 TulipaIO = "7b3808b7-0819-42d4-885c-978ba173db11"
 
 [sources]
-TulipaEnergyModel = {path = ".."}
+TulipaEnergyModel = { path = ".." }
+
+[workspace]
+projects = ["profiling"]
 
 [compat]
 BenchmarkTools = "1"
 DuckDB = "1"
-MetaGraphsNext = "0.7, 0.8"
 TulipaIO = "0.5"

[…]
