### 2026-02-10 8e6d24c
Apply suggestions from code review
Co-authored-by: gnawin <125902905+gnawin@users.noreply.github.com>
