+++ test/test-constraint-storage-balance.jl
@@ -43,7 +43,7 @@
     )
 
     """
-        create_storage_balance_problem(; inflows_profile, period_duration, num_rps)
+        create_storage_balance_problem(; inflows_profile, num_timesteps, num_rps)
 
     Create a storage balance test problem with two storage assets.
     Returns the database connection with configured storage assets and clustering.
@@ -52,7 +52,7 @@
         inflows_profile::Dict{Tuple{String,Int,Int},Vector{Float64}} = Dict(
             ("seasonal_storage", 2030, 1) => [1.0, 5.5, 10.0],
         ),
-        period_duration::Int = 1,
+        num_timesteps::Int = 1,
         num_rps::Int = 2,
     )
         tulipa = TB.TulipaData()
@@ -101,7 +101,7 @@
         layout = TC.ProfilesTableLayout(; cols_to_crossby = [:scenario])
         TC.cluster!(
             connection,
-            period_duration,
+            num_timesteps,
             num_rps;
             method = :convex_hull,
             weight_type = :convex,
@@ -114,7 +114,7 @@
     end
 
     """
-        incoming, outgoing = get_flow_ids(connection, storage_asset, num_rps, period_duration)
+        incoming, outgoing = get_flow_ids(connection, storage_asset, num_rps, num_timesteps)
 
     Query incoming and outgoing flow IDs for a storage asset across all time blocks.
     Returns two dictionaries indexed by (rep_period, time_block_start).
@@ -125,14 +125,14 @@
         connection::DuckDB.DB,
         storage_asset::String,
         num_rps::Int,
-        period_duration::Int,
+        num_timesteps::Int,
     )
         # Pre-allocate dictionaries for all time blocks
         incoming = Dict{Tuple{Int,Int},Vector{Int}}()
         outgoing = Dict{Tuple{Int,Int},Vector{Int}}()
 
         # Initialize empty vectors for each time block
-        for rp in 1:num_rps, tb in 1:period_duration
+        for rp in 1:num_rps, tb in 1:num_timesteps
             incoming[(rp, tb)] = Int[]
             outgoing[(rp, tb)] = Int[]
         end
@@ -243,7 +243,7 @@
         scenario::Int,
         period::Int,
         num_rps::Int,
-        period_duration::Int,
+        num_timesteps::Int,
         config::StorageConfig,
     )
         incoming_expr = JuMP.AffExpr(0.0)
@@ -257,7 +257,7 @@
             weight = get(period_weight_map, (scenario, period, rp), 0.0)
             iszero(weight) && continue  # Skip zero-weight periods
 
-            for tb in 1:period_duration
+            for tb in 1:num_timesteps
                 # Compute weighted charging flows
                 charging_flow = compute_flow_terms(
                     flow,
@@ -328,7 +328,7 @@
     end
 
     """
-        setup_test_problem(storage_asset, inflows_profile, period_duration, num_rps)
+        setup_test_problem(storage_asset, inflows_profile, num_timesteps, num_rps)
 
     Common setup function for creating and configuring test problems.
     Returns connection, energy_problem, flow, profiles, incoming_flow_ids, and outgoing_flow_ids.
@@ -336,13 +336,13 @@
     function setup_test_problem(
         storage_asset::String,
         inflows_profile::Dict{Tuple{String,Int,Int},Vector{Float64}},
-        period_duration::Int,
+        num_timesteps::Int,
         num_rps::Int,
     )
         # Create and configure the test problem
         connection = create_storage_balance_problem(;
             inflows_profile = inflows_profile,
-            period_duration = period_duration,
+            num_timesteps = num_timesteps,
             num_rps = num_rps,
         )
         TEM.populate_with_defaults!(connection)
@@ -358,7 +358,7 @@
 
[…]
