### 2026-02-10 7a2f528
Add test for storage energy balance (#1478)
* Add test for storage energy balance

* Add test for non seasonal storage

* Improve testing

* Update TulipaBuilder version

* Update test with 2 scenarios

* Improve readability, type stability, efficiency

* Update TulipaBuilder version

* Apply suggestions from code review

Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>

* Apply suggestions from code review

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: gnawin <125902905+gnawin@users.no
[…]
