+++ test/test-constraint-storage-balance.jl
@@ -1,0 +1,611 @@
+@testsnippet StorageSetup begin
+    using DuckDB: DuckDB
+    using TulipaBuilder: TulipaBuilder as TB
+    using TulipaClustering: TulipaClustering as TC
+
+    # Type-stable storage configuration struct
+    struct StorageConfig
+        inflows::Float64
+        is_seasonal::Bool
+        initial_units::Float64
+        initial_storage_units::Float64
+        charging_efficiency::Float64
+        discharging_efficiency::Float64
+        initial_storage_level::Float64
+        capacity::Float64
+        capacity_storage_energy::Float64
+    end
+
+    # Storage asset configuration data
+    const STORAGE_CONFIGS = Dict{String,StorageConfig}(
+        "seasonal_storage" => StorageConfig(
+            10.0,  # inflows
+            true,  # is_seasonal
+            1.0,   # initial_units
+            1.0,   # initial_storage_units
+            0.85,  # charging_efficiency
+            0.8,   # discharging_efficiency
+            0.5,   # initial_storage_level
+            1.0,   # capacity
+            168.0, # capacity_storage_energy
+        ),
+        "non_seasonal_storage" => StorageConfig(
+            3.5,   # inflows
+            false, # is_seasonal
+            1.0,   # initial_units
+            1.0,   # initial_storage_units
+            0.9,   # charging_efficiency
+            0.95,  # discharging_efficiency
+            0.25,  # initial_storage_level
+            1.0,   # capacity
+            4.0,   # capacity_storage_energy
+        ),
+    )
+
+    """
+        create_storage_balance_problem(; inflows_profile, num_timesteps, num_rps)
+
+    Create a storage balance test problem with two storage assets.
+    Returns the database connection with configured storage assets and clustering.
+    """
+    function create_storage_balance_problem(;
+        inflows_profile::Dict{Tuple{String,Int,Int},Vector{Float64}} = Dict(
+            ("seasonal_storage", 2030, 1) => [1.0, 5.5, 10.0],
+        ),
+        num_timesteps::Int = 1,
+        num_rps::Int = 2,
+    )
+        tulipa = TB.TulipaData()
+
+        # Add basic producer and consumer for flow balance
+        TB.add_asset!(tulipa, "generator", :producer)
+        TB.add_asset!(tulipa, "consumer", :consumer)
+        TB.add_flow!(tulipa, "generator", "consumer")
+
+        # Add and configure storage assets
+        for (storage_name, config) in STORAGE_CONFIGS
+            TB.add_asset!(
+                tulipa,
+                storage_name,
+                :storage;
+                storage_inflows = config.inflows,
+                is_seasonal = config.is_seasonal,
+                initial_units = config.initial_units,
+                initial_storage_units = config.initial_storage_units,
+                storage_charging_efficiency = config.charging_efficiency,
+                storage_discharging_efficiency = config.discharging_efficiency,
+                initial_storage_level = config.initial_storage_level,
+                capacity = config.capacity,
+                capacity_storage_energy = config.capacity_storage_energy,
+            )
+            TB.add_flow!(tulipa, "consumer", storage_name)
+            TB.add_flow!(tulipa, storage_name, "consumer")
+
+            # Attach inflows profiles per  storage asset, year and scenario
+            for ((asset, year, scenario), values) in inflows_profile
+                if asset == storage_name
+                    TB.attach_profile!(
+                        tulipa,
+                        storage_name,
+                        :inflows,
+                        year,
+                        values;
[…]
