+++ test/Project.toml
@@ -34,6 +34,6 @@ Statistics = "1"
 TOML = "1"
 TestItemRunner = "1"
 TestItems = "1"
-TulipaBuilder = "0.1.2"
+TulipaBuilder = "0.1.4"
 TulipaClustering = "0.5"
 TulipaIO = "0.5"
