### 2026-02-11 9fef637
Fix pkg test example
