### 2026-02-11 d133d1b
Remove asset type "hub" (#1399)
* Remove hub asset from docs.

* Remove hub from model and schema.

* Remove hub from tests, tutorials, and benchmark data.

* Fix formatting

* Reformat with JuliaFormatter@V1

* Update mps files

* Remove hub from OBZ example

* Revert MPS file changes from PR

* Update MPS files (small value changes are from dependency updates)
