### 2026-02-12 06ca101
fixed constraints and added data
