+++ src/constraints/uc-sd-upper-bound-2var.jl
@@ -1,0 +1,142 @@
+export add_shut_down_upper_bound_2var_constraints!
+
+"""
+    add_shut_down_upper_bound_2var_constraints!(model, constraints)
+
+Adds the 2var shut down constraints to the model.
+"""
+function add_shut_down_upper_bound_2var_constraints!(
+    connection,
+    model,
+    variables,
+    expressions,
+    constraints,
+)
+    let table_name = :shut_down_upper_bound_simple_investment_2var, cons = constraints[table_name]
+        expr_avail_simple_method =
+            expressions[:available_asset_units_simple_method].expressions[:assets]
+
+        indices = _append_available_units_start_up_simple_method_2var(connection, table_name)
+
+        units_on_vars = variables[:units_on].container
+        start_up_vars = variables[:start_up].container
+
+        attach_constraint!(
+            model,
+            cons,
+            table_name,
+            [
+                if row.time_block_start == 1
+                    @constraint(model, 0 == 0) # TODO: Placeholder for case k = 1
+
+                # @constraint(
+                #     model,
+                #     start_up_vars[row.start_up_id] <= expr_avail_simple_method[row.avail_id],
+                #     base_name = "$table_name[$(row.asset),$(row.year),$(row.rep_period),$(row.time_block_start):$(row.time_block_end)]"
+                # )
+                else
+                    @constraint(
+                        model,
+                        start_up_vars[row.start_up_id] <=
+                        expr_avail_simple_method[row.avail_id] - units_on_vars[row.units_on_id-1],
+                        base_name = "$table_name[$(row.asset),$(row.year),$(row.rep_period),$(row.time_block_start):$(row.time_block_end)]"
+                    )
+                end for row in indices
+            ],
+        )
+    end
+
+    # let table_name = :shut_down_upper_bound_compact_investment, cons = constraints[table_name]
+    #     expr_avail_compact_method =
+    #         expressions[:available_asset_units_compact_method].expressions[:assets]
+
+    #     indices = _append_available_units_shut_down_compact_method(connection, table_name)
+
+    #     units_on_vars = variables[:units_on].container
+    #     shut_down_vars = variables[:shut_down].container
+
+    #     attach_constraint!(
+    #         model,
+    #         cons,
+    #         table_name,
+    #         [
+    #             @constraint(
+    #                 model,
+    #                 shut_down_vars[row.shut_down_id] <=
+    #                 sum(expr_avail_compact_method[avail_id] for avail_id in row.avail_indices) -
+    #                 units_on_vars[row.units_on_id],
+    #                 base_name = "$table_name[$(row.asset),$(row.year),$(row.rep_period),$(row.time_block_start):$(row.time_block_end)]"
+    #             ) for row in indices
+    #         ],
+    #     )
+    # end
+
+    return nothing
+end
+
+function _append_available_units_start_up_simple_method_2var(connection, table_name)
+    return DuckDB.query(
+        connection,
+        "SELECT
+            cons.*,
+            expr_avail.id AS avail_id,
+            var_units_on.id as units_on_id,
+            var_start_up.id as start_up_id,
+            cons.time_block_start
+        FROM cons_$table_name AS cons
+        LEFT JOIN expr_available_asset_units_simple_method AS expr_avail
+            ON cons.asset = expr_avail.asset
+            AND cons.year = expr_avail.milestone_year
+        LEFT JOIN asset
+            ON cons.asset = asset.asset
+        LEFT JOIN var_units_on
+            ON var_units_on.asset = cons.asset
+            AND var_units_on.year = cons.year
[…]
