### 2026-02-12 b39f996
fixed inclusion/exclusion of default tulipa ramping
