+++ .pre-commit-config.yaml
@@ -42,7 +42,7 @@ repos:
     hooks:
       - id: yamllint
   - repo: https://github.com/domluna/JuliaFormatter.jl
-    rev: v2.2.1
+    rev: v2.3.0
     hooks:
       - id: julia-formatter
         name: Julia Formatter
