### 2026-02-16 2d5d8c7
Fix mixed case of commodity_price profiles (#1492)
Allow commodity_price profiles to be set for one flow and unset for another.

Fixes #1489
