### 2026-02-16 94e858e
Update TulipaClustering requirement to 0.5.1 in /benchmark
Updates the requirements on [TulipaClustering](https://github.com/TulipaEnergy/TulipaClustering.jl) to permit the latest version.
- [Release notes](https://github.com/TulipaEnergy/TulipaClustering.jl/releases)
- [Commits](https://github.com/TulipaEnergy/TulipaClustering.jl/compare/v0.1.0...v0.5.1)

---
updated-dependencies:
- dependency-name: TulipaClustering
  dependency-version: 0.5.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
