### 2026-02-17 066e2fe
Apply suggestion from @abelsiqueira
