+++ benchmark/Project.toml
@@ -15,5 +15,5 @@ projects = ["profiling"]
 [compat]
 BenchmarkTools = "1"
 DuckDB = "1"
-TulipaClustering = "0.5.1"
+TulipaClustering = "0.5"
 TulipaIO = "0.5"
