### 2026-02-17 a8c5b74
Apply suggestion from @abelsiqueira
