+++ benchmark/Project.toml
@@ -15,6 +15,6 @@ projects = ["profiling"]
 [compat]
 BenchmarkTools = "1"
 DuckDB = "1"
-TulipaBuilder = "0.1.5"
+TulipaBuilder = "0.1"
 TulipaClustering = "0.5"
 TulipaIO = "0.5"
