+++ docs/src/10-tutorials/18-stochastic.md
@@ -2,7 +2,7 @@
 
 ## Introduction
 
-Stochastic programming is often used to represent the uncertainty in medium- and long-term optimization problems in energy systems, based on the basis that the *uncertainty can be represented by a known probability distribution*. In the context of energy systems, this uncertainty can be related to the availability of renewable energy sources, demand fluctuations, or hydro inflows. Stochastic programming allows for the incorporation of uncertainty by sampling the uncertainty space and creating multiple scenarios that represent different possible future outcomes. This approach enables decision-makers to make informed choices that are prepared against a range of possible future conditions.
+Stochastic programming is often used to represent the uncertainty in medium- and long-term optimization problems in energy systems, based on the basis that the *uncertainty can be represented by a known probability distribution*. In the context of energy systems, this **uncertainty can be related to the availability of renewable energy sources, demand fluctuations, or hydro inflows**. Stochastic programming allows for the incorporation of uncertainty by sampling the uncertainty space and creating multiple scenarios that represent different possible future outcomes. This approach enables decision-makers to make informed choices that are prepared against a range of possible future conditions.
 
 A **two-stage stochastic** setting is when there are first stage decisions that are unique for each scenario and there are second stage decisions that are made after the uncertainty is realized. In the context of energy systems, the first stage decisions could be related to investment decisions, such as the capacity of new renewable energy sources to be built, while the second stage decisions could be related to operational decisions, such as how to dispatch the available generation resources to meet demand.
 
@@ -14,56 +14,73 @@ TulipaEnergyModel.jl approach to two-stage stochastic optimization is based on t
 
 In the stochastic setting, RPs can be clustered Per or Cross the stochastic scenarios. In that case, the representative periods mapping matrix will relate original periods to representative periods either per scenario (diagonal block structure) or across scenarios (full matrix structure). Here is the concept documentation for more detail: [Clustering per or Cross](https://tulipaenergy.github.io/TulipaClustering.jl/stable/20-concepts/#Clustering-Per-or-Cross)
 
-The TulipaClustering.jl package allows the creation of blended Representative Periods (RPs) to reduce the size of the problem. A tutorial on this package is available under [Tutorial4](https://tulipaenergy.github.io/TulipaEnergyModel.jl/stable/10-tutorials/15-clustering-rep-periods/)
+Let's see how to implement two-stage stochastic optimization with TulipaEnergyModel.jl and TulipaClustering.jl.
 
-## Previously in the TLC
+## Loading the basic input data
 
-We reuse the instantiation from previous tutorials, and subsequently use the data of tutorial 9
+We import the necessary packages, and load the data in the DuckDB.
 
-```julia
-using Pkg: Pkg
-Pkg.activate(".")
-Pkg.instantiate()
-Pkg.add("TulipaClustering")
-Pkg.add("Distances") # only if update packages
+!!! tip "Remember!"
+    1. Activate your local enviroment and instantiate the project, if you haven't already.
+    2. Adjust the path to where you have the data stored.
+
+```@example stochastic
 import TulipaIO as TIO
-import TulipaEnergyModel as TEM #forces TEM. before calling functions
+import TulipaEnergyModel as TEM
 import TulipaClustering as TC
 import CSV
+import Distances
 using DuckDB
 using DataFrames
-import Distances
 using Plots
 using Statistics
 
 connection = DBInterface.connect(DuckDB.DB)
[…]
