### 2026-02-17 e13bb89
Add Tutorial Stochastic Optimization (#1461)
Co-authored-by: Diego Alejandro Tejada Arango <diego.tejadaarango@tno.nl>
