+++ docs/src/10-tutorials/18-stochastic.md
@@ -1,0 +1,316 @@
+# Two-Stage Stochastic Optimization
+
+## Introduction
+
+Stochastic programming is often used to represent the uncertainty in medium- and long-term optimization problems in energy systems, based on the basis that the *uncertainty can be represented by a known probability distribution*. In the context of energy systems, this **uncertainty can be related to the availability of renewable energy sources, demand fluctuations, or hydro inflows**. Stochastic programming allows for the incorporation of uncertainty by sampling the uncertainty space and creating multiple scenarios that represent different possible future outcomes. This approach enables decision-makers to make informed choices that are prepared against a range of possible future conditions.
+
+A **two-stage stochastic** setting is when there are first stage decisions that are unique for each scenario and there are second stage decisions that are made after the uncertainty is realized. In the context of energy systems, the first stage decisions could be related to investment decisions, such as the capacity of new renewable energy sources to be built, while the second stage decisions could be related to operational decisions, such as how to dispatch the available generation resources to meet demand.
+
+![two stage stochastic programming](../figs/two-stage-stochastic-programming.png)
+
+## Two-Stage Stochastic Optimization with TulipaEnergyModel.jl
+
+TulipaEnergyModel.jl approach to two-stage stochastic optimization is based on the concept of Representative Periods (RPs). RPs are a way to reduce the size of the problem by clustering similar time periods together, see tutorial [Blended Representative Periods with Tulipa Clustering](@ref blended-representative-periods). This is particularly useful in the context of energy systems, where there can be a large number of time periods to consider. By clustering similar time periods together, we can reduce the number of variables and constraints in the optimization problem, making it more tractable.
+
+In the stochastic setting, RPs can be clustered Per or Cross the stochastic scenarios. In that case, the representative periods mapping matrix will relate original periods to representative periods either per scenario (diagonal block structure) or across scenarios (full matrix structure). Here is the concept documentation for more detail: [Clustering per or Cross](https://tulipaenergy.github.io/TulipaClustering.jl/stable/20-concepts/#Clustering-Per-or-Cross)
+
+Let's see how to implement two-stage stochastic optimization with TulipaEnergyModel.jl and TulipaClustering.jl.
+
+## Loading the basic input data
+
+We import the necessary packages, and load the data in the DuckDB.
+
+!!! tip "Remember!"
+    1. Activate your local enviroment and instantiate the project, if you haven't already.
+    2. Adjust the path to where you have the data stored.
+
+```@example stochastic
+import TulipaIO as TIO
+import TulipaEnergyModel as TEM
+import TulipaClustering as TC
+import CSV
+import Distances
+using DuckDB
+using DataFrames
+using Plots
+using Statistics
+
+connection = DBInterface.connect(DuckDB.DB)
+input_dir = "my-awesome-energy-system/tutorial-9"
+TIO.read_csv_folder(connection, input_dir)
+```
+
+Here is a schematic of the energy system we are modeling:
+
+```mermaid
+flowchart TD
+ens            --> e_demand
+ocgt           --> e_demand
+ccgt           --> e_demand
+wind           --> e_demand
+solar          --> e_demand
+electrolizer   --> h2_demand
+e_demand       --> electrolizer
+smr_ccs        --> h2_demand
+wind_offshore  --> e_demand
+wind_offshore  --> electrolizer
+battery        --> e_demand
+e_demand       --> battery
+h2_storage     --> h2_demand
+h2_demand      --> h2_storage
+hydro_reservoir--> e_demand
+hydro_reservoir--> spillage
+```
+
+## Profiles Depending on the Scenario
+
[…]
