### 2026-02-17 a939e2b
Update TulipaBuilder requirement to 0.1.5 in /benchmark (#1493)
* Update TulipaBuilder requirement to 0.1.5 in /benchmark

Updates the requirements on [TulipaBuilder](https://github.com/TulipaEnergy/TulipaBuilder.jl) to permit the latest version.
- [Release notes](https://github.com/TulipaEnergy/TulipaBuilder.jl/releases)
- [Commits](https://github.com/TulipaEnergy/TulipaBuilder.jl/compare/v0.1.0-trigger-docs...v0.1.5)

---
updated-dependencies:
- dependency-name: TulipaBuilder
  dependency-version: 0.1.5
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.
[…]
