### 2026-02-19 a047a8f
TLC workflow tutorial squashed (#1405)
* TLC workflow tutorial squashed

* Fix lint

* Update file names

* Update of tutorial after rebase

---------

Co-authored-by: Diego Alejandro Tejada Arango <diego.tejadaarango@tno.nl>
