+++ docs/src/10-tutorials/20-workflow-from-raw-data.md
@@ -1,0 +1,777 @@
+# [Data-Handling Workflow From Raw Data (Example)](@id workflow-tutorial-raw-data)
+
+In previous tutorials you have learned how to work with TulipaEnergyModel and related tools such as TulipaClustering and TulipaIO. However, we have not yet considered data requirements and processing to get to these steps. How do we work with 'raw' data and convert it to the format needed for running the optimization model? How are results stored? How do we compare scenarios with different assumptions for input data? This tutorial will be about teaching you the basics in the workflow that are currently supported.
+
+!!! note "Important note"
+    There are many ways to go about processing data, depending on your preferences. The work below is simply an example! For a general overview please visit the [Analysis Workflow](@ref data) section.
+
+## 1. Set up
+
+!!! warning "Version compatibility"
+    The code in this tutorial is based on TulipaEnergyModel version 0.18.2. If you are on a different version, some of the code may not work as expected. We recommend creating a separate environment for this tutorial. You can check your version and update if necessary using the following code:
+
+```julia
+# Check what TulipaEnergyModel version you are on
+Pkg.status()
+# or in console type ] followed by 'status TulipaEnergyModel'
+# if necessary run the following to ensure the compatible version for this tutorial is installed:
+# Pkg.add(name="TulipaEnergyModel", version="0.18.2")
+```
+
+Like previous tutorials we will start with working in our 'workflow.jl' file. There are a couple of new packages we will add first.
+
+```julia
+# Guarantee to run in the current directory
+using Pkg: Pkg
+Pkg.activate(".")
+
+# Add new packages specific for this tutorial
+Pkg.add("CSV")          # For working with CSV files
+Pkg.add("Chain")        # For handling data transformation
+Pkg.add("Tables")       # For manipulating the data tables
+Pkg.add("DBInterface")  # For conducting queries
+Pkg.add("ODBC")         # For working with Microsoft Access database files
+Pkg.add("DataFrames")
+Pkg.add("Plots")
+
+# Load the packages
+import TulipaIO as TIO
+import TulipaEnergyModel as TEM
+using DuckDB
+using DBInterface
+using DataFrames
+using Plots
+using CSV
+using Chain
+using Tables
+using ODBC
+```
+
+## 2. Context and input data
+
+We are interested in investigating the electricity prices in our future system, in moments where both solar PV and wind make up a large share of our electricity production. This will help us understand the role that dispatchable/storage technologies play within the system.
+
+All input data required for this tutorial can be found here: <https://zenodo.org/records/17454936>
+
+Download the data and place them in a directory similar to the previous tutorials (within your VS project), for example `my-awesome-energy-system/tutorial-7`. As can be seen in this folder, the data is comprised of different file types. We have .csv files for our hourly profiles, .xlsx files for our asset and technology data, and a .accdb file for storage metadata.
+
+The first thing we should do is transforming and cleaning the data. As you know by now, TulipaEnergyModel requires input data to be in a specific format following the schemas as explained in the documentation.
+
+## 3. Transforming and cleaning the data
+
[…]
