### 2026-02-23 df96aba
Add validation for consistency between investable and asset_both
Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>
