### 2026-02-23 b7512d3
Apply JuliaFormatter to fix lint failures
Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>
