### 2026-02-24 6498a58
Add some dev docs for Aqua and Aqua badge on README
