### 2026-02-24 6f34d69
Add Aqua tests and fix the issues found (#1498)
* Add Aqua tests and fix the issues found

Adds Aqua.test_all from Aqua.jl.
Aqua.jl provides some quality checks, like wrong exports and type piracies.
This commit also fixes the issues found at the moment.

* Add some dev docs for Aqua and Aqua badge on README
