+++ test/test-aqua.jl
@@ -1,0 +1,4 @@
+@testitem "Aqua.jl" setup = [CommonSetup] begin
+    using Aqua
+    Aqua.test_all(TulipaEnergyModel)
+end
