+++ Project.toml
@@ -21,7 +21,6 @@ DuckDB_jll = "2cbbab25-fc8b-58cf-88d4-687a02676033"
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 ParametricOptInterface = "0ce4ce61-57bf-432b-a095-efac525d185e"
