### 2026-02-24 9b0ecfb
Apply suggestion from @abelsiqueira
Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>
