+++ test/test-data-validation.jl
@@ -931,8 +931,10 @@ end
         :milestone_year => [1, 1, 1],
         :investable => [true, true, false],
     )
+    # A is missing and should generated an error message.
+    # C is missing as well, but is not investable, so it is fine.
     asset_both =
-        DataFrame(:asset => ["B", "C"], :milestone_year => [1, 1], :commission_year => [1, 1])
+        DataFrame(:asset => ["B"], :milestone_year => [1], :commission_year => [1])
     DuckDB.register_data_frame(connection, asset_milestone, "asset_milestone")
     DuckDB.register_data_frame(connection, asset_both, "asset_both")
 
