### 2026-02-24 2a205c8
[Validation] Add consistency check between `investable` and `asset_both` (#1497)
* Initial plan

* Add validation for consistency between investable and asset_both

Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>

* Apply JuliaFormatter to fix lint failures

Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>

* Fix Lint for copilot :)

* Apply suggestion from @abelsiqueira

Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>

* Fix lint

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-
[…]
