+++ test/test-constraint-storage-balance.jl
@@ -98,7 +98,7 @@
 
         # Create connection and apply clustering
         connection = TB.create_connection(tulipa)
-        layout = TC.ProfilesTableLayout(; cols_to_crossby = [:scenario])
+        layout = TC.ProfilesTableLayout(; year = :milestone_year, cols_to_crossby = [:scenario])
         TC.cluster!(
             connection,
             num_timesteps,
