+++ docker-compose.yml
@@ -1,0 +1,10 @@
+services:
+  tulipa:
+    image: julia:1.10
+    container_name: tulipa-dev
+    working_dir: /workspace
+    volumes:
+      - ./:/workspace
+    stdin_open: true
+    tty: true
+    command: sleep infinity
