### 2026-02-27 c22d972
resource allocation limits
