+++ test/Project.toml
@@ -35,6 +35,6 @@ Statistics = "1"
 TOML = "1"
 TestItemRunner = "1"
 TestItems = "1"
-TulipaBuilder = "0.1.4"
+TulipaBuilder = "0.2"
 TulipaClustering = "0.5"
 TulipaIO = "0.5"
