### 2026-03-02 637aec5
Update TulipaBuilder requirement from 0.1.4 to 0.1.4, 0.2 in /docs
Updates the requirements on [TulipaBuilder](https://github.com/TulipaEnergy/TulipaBuilder.jl) to permit the latest version.
- [Release notes](https://github.com/TulipaEnergy/TulipaBuilder.jl/releases)
- [Commits](https://github.com/TulipaEnergy/TulipaBuilder.jl/compare/v0.1.4...v0.2.0)

---
updated-dependencies:
- dependency-name: TulipaBuilder
  dependency-version: 0.2.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
