+++ docs/Project.toml
@@ -32,6 +32,6 @@ JuMP = "1"
 LiveServer = "1"
 OrderedCollections = "1"
 Plots = "1"
-TulipaBuilder = "0.1.4"
+TulipaBuilder = "0.1.4, 0.2"
 TulipaClustering = "0.4 - 0.5"
 TulipaIO = "0.5.0"
