### 2026-03-02 e6ed025
Update Aqua requirement to 0.8.14 in /test
Updates the requirements on [Aqua](https://github.com/JuliaTesting/Aqua.jl) to permit the latest version.
- [Release notes](https://github.com/JuliaTesting/Aqua.jl/releases)
- [Changelog](https://github.com/JuliaTesting/Aqua.jl/blob/master/CHANGELOG.md)
- [Commits](https://github.com/JuliaTesting/Aqua.jl/compare/v0.1.0...v0.8.14)

---
updated-dependencies:
- dependency-name: Aqua
  dependency-version: 0.8.14
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
