### 2026-03-03 df2d0c1
Update utils/scripts/split-graph-data.jl
Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>
