### 2026-03-03 82b514f
year -> milestone_year (#1502)
* year -> milestone_year

* Add custom key to TC

* Ensure TB is compatible with this PR

* Remove stale dependencies

* Change TB version in project.toml

* More places

* Update utils/scripts/split-graph-data.jl

Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>

---------

Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>
