+++ src/model-preparation.jl
@@ -18,8 +18,8 @@ Computes the incoming and outgoing expressions per row of `cons` for the constra
 that are within the representative periods.
 
 Include_commission_year is only used for the constraints regarding the flows for the semi-compact investment method.
-If true, the expression will include the commission year of the flows.
-If false (the default), it will only use the year (i.e., milestone_year).
+If true, the expression will include the commission_year of the flows.
+If false (the default), it will only use the milestone_year.
 
 This function is only used internally in the model.
 
@@ -38,7 +38,7 @@ time block.
 The algorithm works like this:
 
 ```
-1. Loop over each group of (asset, year, rep_period)
+1. Loop over each group of (asset, milestone_year, rep_period)
 1.1. Loop over each variable in the group: (var_id, var_time_block_start, var_time_block_end)
 1.1.1. Loop over each timestep in var_time_block_start:var_time_block_end
 1.1.1.1. Compute the coefficient of the variable based on the rep_period resolution and the variable coefficients
@@ -81,7 +81,7 @@ function add_expression_terms_rep_period_constraints!(
     year_columns = if include_commission_year
         [:milestone_year, :commission_year]
     else
-        [:year]
+        [:milestone_year]
     end
 
     # The SQL strategy to improve looping over the groups and then the
@@ -126,10 +126,10 @@ function add_expression_terms_rep_period_constraints!(
 
         year_select = join(("cons.$col" for col in year_columns), ", ")
         year_join_var = join(("cons.$col = var.$col" for col in year_columns), " AND ")
-        year_join_rp = "cons.$(year_columns[1]) = rep_periods_data.year"
+        year_join_rp = "cons.$(year_columns[1]) = rep_periods_data.milestone_year"
 
         # Start of the algorithm
-        # 1. Loop over each group of (asset, year, rep_period)
+        # 1. Loop over each group of (asset, milestone_year, rep_period)
         for group_row in DuckDB.query(
             connection,
             "SELECT
@@ -288,7 +288,7 @@ function add_expression_terms_over_clustered_year_constraints!(
         connection,
         cons.table_name,
         grouped_cons_table_name,
-        [:asset, :year, :scenario],
+        [:asset, :milestone_year, :scenario],
         [:id, :period_block_start, :period_block_end],
     )
 
@@ -297,7 +297,7 @@ function add_expression_terms_over_clustered_year_constraints!(
         connection,
         "rep_periods_mapping",
         grouped_rpmap_over_rp_table_name,
-        [:year, :scenario, :rep_period],
+        [:milestone_year, :scenario, :rep_period],
         [:period, :weight];
         order_agg_by = :period,
     )
@@ -319,7 +319,7 @@ function add_expression_terms_over_clustered_year_constraints!(
             "CREATE OR REPLACE TEMP TABLE t_groups AS
             SELECT
                 cons.asset,
-                cons.year,
+                cons.milestone_year,
                 cons.scenario,
                 ANY_VALUE(cons.id) AS cons_id_vec,
                 ANY_VALUE(cons.period_block_start) AS cons_period_block_start_vec,
@@ -334,18 +334,18 @@ function add_expression_terms_over_clustered_year_constraints!(
             FROM $grouped_cons_table_name AS cons
             LEFT JOIN $grouped_var_table_name AS var
                 ON cons.asset = var.asset
-                AND cons.year = var.year
+                AND cons.milestone_year = var.milestone_year
             LEFT JOIN $grouped_rpmap_over_rp_table_name AS rpmap
-                ON rpmap.year = cons.year
+                ON rpmap.milestone_year = cons.milestone_year
                 AND rpmap.scenario = cons.scenario
                 AND rpmap.rep_period = var.rep_period
             LEFT JOIN rep_periods_data AS rpdata
-                ON rpdata.year = cons.year
[…]
