+++ test/utils.jl
@@ -230,10 +230,10 @@ end
 
         # Basic time information
         "year_data" => DataFrame(:year => [2030]),
-        "rep_periods_data" => DataFrame(:year => [2030, 2030], :rep_period => [1, 2]),
-        "timeframe_data" => DataFrame(:year => 2030, :period => 1:365),
+        "rep_periods_data" => DataFrame(:milestone_year => [2030, 2030], :rep_period => [1, 2]),
+        "timeframe_data" => DataFrame(:milestone_year => 2030, :period => 1:365),
         "rep_periods_mapping" => DataFrame(
-            :year => 2030,
+            :milestone_year => 2030,
             :scenario => Int32(1),
             :period => 1:365,
             :rep_period => mod1.(1:365, 2),
