+++ src/structures.jl
@@ -225,10 +225,10 @@ Structure to hold the dictionaries of profiles.
 mutable struct ProfileLookup
     # The integers here are Int32 because they are obtained directly from DuckDB
 
-    # rep_period[(asset, year, rep_period)]
+    # rep_period[(asset, milestone_year, rep_period)]
     rep_period::Dict{Tuple{String,Int32,Int32},ProfileWithRollingHorizon}
 
-    # over_clustered_year[(asset, year, scenario)]
+    # over_clustered_year[(asset, milestone_year, scenario)]
     over_clustered_year::Dict{Tuple{String,Int32,Int32},Vector{Float64}}
 end
 
