+++ .github/workflows/Benchmark.yml
@@ -78,7 +78,7 @@ jobs:
             --rev="$BASE,$HEAD" \
             --npart=10 --format=png --input-dir=results/ --output-dir=plots/
       - name: Upload plot as artifact
-        uses: actions/upload-artifact@v6
+        uses: actions/upload-artifact@v7
         with:
           name: plots
           path: plots
