### 2026-03-04 e12762c
Disable pushing preview when the PR is made by dependabot (#1510)
Closes #1509
