### 2026-03-04 d9faedd
Prevent dependabot from trying to comment on PR (#1512)
Related to #1509 and #1504
