+++ .github/workflows/Docs.yml
@@ -65,7 +65,7 @@ jobs:
           comment-author: "github-actions[bot]"
           body-includes: Check the documentation preview
       - name: Comment on PR
-        if: github.event.pull_request.head.repo.full_name == github.repository && steps.docpreview.outputs.comment-id == 0
+        if: github.event.pull_request.head.repo.full_name == github.repository && github.actor != 'dependabot[bot]' && steps.docpreview.outputs.comment-id == 0
         uses: peter-evans/create-or-update-comment@v5
         with:
           comment-id: ${{ steps.docpreview.outputs.comment-id }}
