### 2026-03-04 dfe6241
Update ParametricOptInterface requirement from 0.12.1, 0.14 to 0.12.1, 0.14, 0.15 (#1504)
Updates the requirements on [ParametricOptInterface](https://github.com/jump-dev/ParametricOptInterface.jl) to permit the latest version.
- [Release notes](https://github.com/jump-dev/ParametricOptInterface.jl/releases)
- [Commits](https://github.com/jump-dev/ParametricOptInterface.jl/commits/v0.15.1)

---
updated-dependencies:
- dependency-name: ParametricOptInterface
  dependency-version: 0.15.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by:
[…]
