+++ Project.toml
@@ -36,7 +36,7 @@ JSON = "0.21.4, 1"
 JuMP = "1"
 MathOptInterface = "1"
 OrderedCollections = "1"
-ParametricOptInterface = "0.12.1, 0.14"
+ParametricOptInterface = "0.12.1, 0.14, 0.15"
 Statistics = "1"
 TOML = "1"
 TimerOutputs = "0.5"
