+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.21",
   "v0.20",
   "v0.19",
   "v0.18",
@@ -22,5 +23,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.20.1";
+var DOCUMENTER_NEWEST = "v0.21.0";
 var DOCUMENTER_STABLE = "stable";
