### 2026-03-09 1853fe5
chore: :robot: pre-commit update (#1514)
Co-authored-by: abelsiqueira <1068752+abelsiqueira@users.noreply.github.com>
