+++ .github/workflows/Lint.yml
@@ -25,7 +25,7 @@ jobs:
         with:
           version: "1"
       - name: Use Julia cache
-        uses: julia-actions/cache@v2
+        uses: julia-actions/cache@v3
       - name: Install JuliaFormatter.jl
         run: |
           julia -e 'using Pkg; pkg"add JuliaFormatter@1"' # Version 1 is fixed due to https://github.com/domluna/JuliaFormatter.jl/issues/878
