+++ .github/workflows/Benchmark.yml
@@ -45,7 +45,7 @@ jobs:
       - uses: julia-actions/setup-julia@latest
         with:
           version: 1
-      - uses: julia-actions/cache@v2
+      - uses: julia-actions/cache@v3
       - name: Extract Package Name from Project.toml
         id: extract-package-name
         run: |
