+++ .github/workflows/CompareMPS.yml
@@ -49,7 +49,7 @@ jobs:
         with:
           version: "1"
       - name: Use Julia cache
-        uses: julia-actions/cache@v2
+        uses: julia-actions/cache@v3
       - name: Install the packages
         uses: julia-actions/julia-buildpkg@v1
       - name: Run comparison
