### 2026-03-11 e7f4730
Apply suggestions from code review
Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>
Co-authored-by: Diego Alejandro Tejada Arango <12887482+datejada@users.noreply.github.com>
