+++ debugging/debug.jl
@@ -3,13 +3,14 @@ using DuckDB
 using TulipaIO
 
 in_folder = joinpath(pwd(), "test/inputs")
-# in_folder = joinpath(pwd(), "debugging/experiment-inputs")
+in_folder = joinpath(pwd(), "debugging/experiment-inputs")
 bench_folder = joinpath(pwd(), "debugging")
 
 # in_dir = joinpath(in_folder, "Multi-year Investments")
 # in_dir = joinpath(in_folder, "with_limits_model_2")
 # in_dir = joinpath(in_folder, "all-susd")
-in_dir = joinpath(in_folder, "UC-ramping")
+# in_dir = joinpath(in_folder, "UC-ramping")
+in_dir = joinpath(in_folder, "trajectories-feas")
 # in_dir = joinpath(bench_folder, "EU")
 
 # Conversion
