+++ src/structures.jl
@@ -228,8 +228,8 @@ mutable struct ProfileLookup
     # rep_period[(asset, milestone_year, rep_period)]
     rep_period::Dict{Tuple{String,Int32,Int32},ProfileWithRollingHorizon}
 
-    # over_clustered_year[(asset, milestone_year, scenario)]
-    over_clustered_year::Dict{Tuple{String,Int32,Int32},Vector{Float64}}
+    # inter_period[(asset, milestone_year, scenario)]
+    inter_period::Dict{Tuple{String,Int32,Int32},Vector{Float64}}
 end
 
 """
@@ -244,7 +244,7 @@ It hides the complexity behind the energy problem, making the usage more friendl
 - `variables`: A dictionary of [TulipaVariable](@ref TulipaVariable)s containing the variables of the model.
 - `expressions`: A dictionary of [TulipaExpression](@ref TulipaExpression)s containing the expressions of the model attached to tables.
 - `constraints`: A dictionary of [TulipaConstraint](@ref TulipaConstraint)s containing the constraints of the model.
-- `profiles`: Holds the profiles per `rep_period` or `over_clustered_year` in dictionary format. See [ProfileLookup](@ref).
+- `profiles`: Holds the profiles per `rep_period` or `inter_period` in dictionary format. See [ProfileLookup](@ref).
 - `model_parameters`: A [ModelParameters](@ref ModelParameters) structure to store all the parameters that are exclusive of the model.
 - `model`: A JuMP.Model object representing the optimization model.
 - `solved`: A boolean indicating whether the `model` has been solved or not.
