+++ src/constraints/create.jl
@@ -23,9 +23,9 @@ function compute_constraints_indices(connection)
             :max_ramp_with_unit_commitment,
             :max_ramp_without_unit_commitment,
             :balance_storage_rep_period,
-            :balance_storage_over_clustered_year,
-            :min_energy_over_clustered_year,
-            :max_energy_over_clustered_year,
+            :balance_storage_inter_period,
+            :min_energy_inter_period,
+            :max_energy_inter_period,
             :transport_flow_limit_simple_method,
             :min_outgoing_flow_for_transport_flows_without_unit_commitment,
             :min_outgoing_flow_for_transport_vintage_flows,
