### 2026-03-13 1008070
Update the working example of checking infeasibility through JuMP IIS functionality (#1532)
