### 2026-03-13 a19c020
Apply suggestions from code review
Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>
Co-authored-by: Diego Alejandro Tejada Arango <12887482+datejada@users.noreply.github.com>
