### 2026-03-13 48a1e22
Add copilot instructions for issues and reviews (#1516)
* Add copilot instructions for issues and reviews

* Change to AGENTS.md file

* Apply suggestions from code review

Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>
Co-authored-by: Diego Alejandro Tejada Arango <12887482+datejada@users.noreply.github.com>

* Fix Lint

* Apply suggestions from code review

Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>
Co-authored-by: Diego Alejandro Tejada Arango <12887482+datejada@users.noreply.github.com>

* Add info for developers

* Apply suggestions from code review
[…]
