+++ docs/src/20-user-guide/55-outputs.md
@@ -39,7 +39,7 @@ Each output table has three types of columns:
 
 ### `var_assets_decommission_energy`
 
-For a storage asset that has `storage_method_energy` commissioned in `commission_year`, the optimal decommissioning (decrease) of asset capacity storage energy in `milestone_year`, expressed in the same units as `capacity_storage_energy` of asset.
+For a storage asset with `storage_method_energy = 'optimize_storage_capacity'` commissioned in `commission_year`, the optimal decommissioning (decrease) of asset capacity storage energy in `milestone_year`, expressed in the same units as `capacity_storage_energy` of asset.
 
 Associated input parameters: `investment_integer_storage_energy`, `capacity_storage_energy`
 
@@ -51,7 +51,7 @@ Associated input parameters: `decommissionable`, `initial_units`, `investment_in
 
 ### `var_assets_investment_energy`
 
-For a storage asset that has `storage_method_energy`, the optimal investment (increase) in asset capacity storage energy in `milestone_year`, expressed in the same units as `capacity_storage_energy` of asset.
+For a storage asset with `storage_method_energy = 'optimize_storage_capacity'`, the optimal investment (increase) in asset capacity storage energy in `milestone_year`, expressed in the same units as `capacity_storage_energy` of asset.
 
 Associated input parameters: `investable`, `investment_integer_storage_energy`, `capacity_storage_energy`, `investment_limit_storage_energy`
 
