+++ previews/PR1533/10-tutorials/my-awesome-energy-system/tutorial-6-compact-method/model-parameters-example.toml
@@ -1,0 +1,3 @@
+discount_rate = 0.03
+discount_year = 2020
+power_system_base = 100
