### 2026-03-16 5a88d21
Add extra flows for trajectories
