### 2026-03-16 082aa48
moved trajectories to separate flows
