### 2026-03-16 f32b755
Apply suggestions from code review
Remove dead comments, thanks github AI

Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
