### 2026-03-16 4d946ba
Add tests for risk aversion parameters
