### 2026-03-16 5593351
Add model parameters for risk aversion: weight lambda and confidence level alpha (#1534)
* Add model parameters for risk aversion: weight lambda and confidence level alpha

* Add tests for risk aversion parameters
