### 2026-03-16 c7f3e70
Add variable Value at Risk threshold mu and tests
