### 2026-03-16 3fe5424
docs: update docs/src/index.md
