### 2026-03-16 aaca430
Create table and save outputs
Co-Authored-By: Claude Code (claude-sonnet-4-6) <noreply@anthropic.com>
