### 2026-03-16 e6a0f86
docs: add florinelaseur as a contributor for code, research, and doc (#1536)
* docs: update README.md

* docs: update docs/src/index.md

* docs: update .all-contributorsrc

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
