### 2026-03-16 c5aaeae
Refactor to use expressions
Co-Authored-By: Claude Code (claude-sonnet-4-6) <noreply@anthropic.com>
