### 2026-03-17 2b652c8
Apply suggestion from @abelsiqueira
