### 2026-03-17 353c420
Remove model_parameters_file from the code
