### 2026-03-17 ef819ac
Add tail excess slack variable xi and tests
