### 2026-03-17 a0acc16
Refactor model parameters struct
