### 2026-03-17 2739dea
Update functions with latest parameters
