### 2026-03-17 b5937bb
Add CVaR term to objective and tests
