### 2026-03-17 b474c7a
Fix printing test
