+++ test/test-io.jl
@@ -59,13 +59,11 @@ end
 
     io = IOBuffer()
     TulipaEnergyModel.solve_model!(energy_problem)
+    TulipaEnergyModel.save_solution!(energy_problem; compute_duals = false)
     print(io, energy_problem)
     expected_output = split(read("io-outputs/energy-problem-model-solved.txt", String))
     actual_output = split(String(take!(io)))
-    # To make the comparison more robust, the only look at the the first
-    # 9 significant digits of the solution (+ '.')
-    expected_output[end] = expected_output[end][1:10]
-    actual_output[end] = actual_output[end][1:10]
 
-    @test expected_output == actual_output
+    @test length(expected_output) == length(actual_output)
+    @test all(compare_tokens.(expected_output, actual_output))
 end
