### 2026-03-17 aa6d403
Only print when there is a valid objective value
