### 2026-03-18 e730cb0
Apply suggestions from code review
Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>
