### 2026-03-18 886c351
Apply suggestions from code review
