### 2026-03-18 cd3b0de
Refactor model parameters (#1542)
* Update input schema with new table

* Allow new table to be missing

* Remove model_parameters_file from the code

* Remove old TOML files

* Add new tables to the relevant case studies

* Refactor model parameters struct

* Update tests

* Update docs

* Update functions with latest parameters

* Update model parameter files

* MPS file (intended)

* Apply suggestions from code review

Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>

* Apply suggestions from code review

---------

Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail
[…]
