+++ src/model-parameters.jl
@@ -1,67 +1,100 @@
 export ModelParameters
 
 """
-    ModelParameters(;key = value, ...)
-    ModelParameters(path; ...)
-    ModelParameters(connection; ...)
-    ModelParameters(connection, path; ...)
+    ModelParameters(connection)
 
 Structure to hold the model parameters.
-Some values are defined by default and some required explicit definition.
 
-If `path` is passed, it is expected to be a string pointing to a TOML file with
-a `key = value` list of parameters. Explicit keyword arguments take precedence.
+Create parameters by reading from the `model_parameters` table in the DuckDB
+connection. Default values for `discount_rate` and `power_system_base` come from
+the input schema. The `discount_year` is always computed as the minimum between
+the provided discount year (table value if present, otherwise schema default) and
+the minimum year in `year_data` where `is_milestone` is true.
 
-If `connection` is passed, the default `discount_year` is set to the
-minimum of all milestone years. In other words, we check for the table
-`year_data` for the column `year` where the column `is_milestone` is true.
-Explicit keyword arguments take precedence.
+## Fields
 
-If both are passed, then `path` has preference. Explicit keyword arguments take precedence.
-
-## Parameters
-
-- `discount_rate::Float64 = 0.0`: The model discount rate.
+- `discount_rate::Float64`: The model discount rate.
 - `discount_year::Int`: The model discount year.
-- `power_system_base::Float64 = 100.0`: The power system base in MVA.
-- `risk_aversion_weight_lambda::Float64 = 0.0`: Weight reflecting the risk aversion of the objective
-- `risk_aversion_confidence_level_alpha::Float64 = 0.95`: Confidence level that system costs will not exceed the VaR_alpha
+- `power_system_base::Float64`: The power system base in MVA.
+- `risk_aversion_weight_lambda::Float64`: Weight reflecting the risk aversion of the objective
+- `risk_aversion_confidence_level_alpha::Float64`: Confidence level that system costs will not exceed the VaR_alpha
 """
-Base.@kwdef mutable struct ModelParameters
-    discount_rate::Float64 = 0.0
-    discount_year::Int # Explicit definition expected
-    power_system_base::Float64 = 100.0
-    risk_aversion_weight_lambda::Float64 = 0.0
-    risk_aversion_confidence_level_alpha::Float64 = 0.95
+mutable struct ModelParameters
+    discount_rate::Float64
+    discount_year::Int32
+    power_system_base::Float64
+    risk_aversion_weight_lambda::Float64
+    risk_aversion_confidence_level_alpha::Float64
 end
 
-# Using `@kwdef` defines a default constructor based on keywords
+function _default_discount_year(connection::DuckDB.DB)
+    return minimum(
+        row.year for row in DuckDB.query(
+            connection,
+            "SELECT year::INT32 AS year FROM year_data WHERE is_milestone = true",
+        )
+    )
+end
 
-function _read_model_parameters(path)
-    if length(path) > 0 && !isfile(path)
-        throw(ArgumentError("path `$path` does not contain a file"))
+function _read_model_parameters(connection::DuckDB.DB)
+    if !_check_if_table_exists(connection, "model_parameters")
+        return Dict{Symbol,Union{Float64,Int32,Int64}}()
     end
 
-    file_data = length(path) > 0 ? TOML.parsefile(path) : Dict{String,Any}()
-    file_parameters = Dict(Symbol(k) => v for (k, v) in file_data)
+    rows = collect(DuckDB.query(
+        connection,
+        "SELECT *
+        FROM model_parameters",
+    ))
 
-    return file_parameters
+    if length(rows) > 1
+        error("Table `model_parameters` must contain at most one row.")
+    end
+
+    if isempty(rows)
[…]
