+++ test/test-model-parameters.jl
@@ -1,5 +1,4 @@
 @testsnippet ModelParametersSetup begin
-    const TOML_PATH = joinpath(@__DIR__, "inputs", "model-parameters-example.toml")
     const NORSE_PATH = joinpath(@__DIR__, "inputs", "Norse")
 
     function connection_with_norse()
@@ -8,69 +7,192 @@
         return connection
     end
 
-    function validate_toml_matches_mp(mp, toml_path)
-        data = TOML.parsefile(toml_path)
-        for (key, value) in data
-            @test value == getfield(mp, Symbol(key))
+    function create_model_parameters_table!(
+        connection;
+        discount_rate = missing,
+        discount_year = missing,
+        power_system_base = missing,
+        risk_aversion_confidence_level_alpha = missing,
+        risk_aversion_weight_lambda = missing,
+    )
+        DuckDB.query(
+            connection,
+            """
+            CREATE OR REPLACE TABLE model_parameters (
+                discount_rate DOUBLE,
+                discount_year INTEGER,
+                power_system_base DOUBLE,
+                risk_aversion_confidence_level_alpha DOUBLE,
+                risk_aversion_weight_lambda DOUBLE
+            );
+            """,
+        )
+
+        # Build a data frame with non-missing values only
+        data_dict = Dict{Symbol,Vector}()
+        !ismissing(discount_rate) && (data_dict[:discount_rate] = [discount_rate])
+        !ismissing(discount_year) && (data_dict[:discount_year] = [discount_year])
+        !ismissing(power_system_base) && (data_dict[:power_system_base] = [power_system_base])
+        !ismissing(risk_aversion_confidence_level_alpha) && (
+            data_dict[:risk_aversion_confidence_level_alpha] =
+                [risk_aversion_confidence_level_alpha]
+        )
+        !ismissing(risk_aversion_weight_lambda) &&
+            (data_dict[:risk_aversion_weight_lambda] = [risk_aversion_weight_lambda])
+
+        if !isempty(data_dict)
+            DuckDB.register_data_frame(connection, DataFrame(data_dict), "model_parameters_values")
+
+            # Build column list dynamically for the INSERT statement
+            cols = join(String.(keys(data_dict)), ", ")
+            DuckDB.query(
+                connection,
+                """
+                INSERT INTO model_parameters($cols)
+                SELECT $cols
+                FROM model_parameters_values
+                """,
+            )
         end
+
+        return
     end
 end
 
-@testitem "Test model parameters - basic usage" setup = [CommonSetup] tags =
-    [:unit, :validation, :fast] begin
-    mp = TulipaEnergyModel.ModelParameters(;
-        discount_rate = 0.1,
-        discount_year = 2018,
-        power_system_base = 50,
+@testitem "Test model parameters - read from model_parameters table" setup =
+    [CommonSetup, ModelParametersSetup] tags = [:unit, :validation, :fast] begin
+    connection = connection_with_norse()
+    create_model_parameters_table!(
+        connection;
+        discount_rate = 0.03,
+        discount_year = 2020,
+        power_system_base = 90.0,
         risk_aversion_confidence_level_alpha = 0.90,
         risk_aversion_weight_lambda = 0.1,
     )
-    @test mp.discount_rate == 0.1
-    @test mp.discount_year == 2018
-    @test mp.power_system_base == 50
+
+    mp = TulipaEnergyModel.ModelParameters(connection)
+
+    @test mp.discount_rate == 0.03
[…]
