### 2026-03-19 c608389
Apply suggestions from code review
Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>
