+++ previews/PR1537/10-tutorials/my-awesome-energy-system/tutorial-6-simple-method/model-parameters-example.toml
@@ -1,3 +1,0 @@
-discount_rate = 0.03
-discount_year = 2020
-power_system_base = 100
