+++ src/solve-model.jl
@@ -111,10 +111,8 @@ function save_solution!(connection, model, variables, constraints; compute_duals
             error("Table obj_breakdown has no corresponding objective in model")
         end
     end
-    breakdown_table = [
-        (name = name, value = JuMP.value(model[Symbol(name)])) for
-        name in breakdown_names
-    ]
+    breakdown_table =
+        [(name = name, value = JuMP.value(model[Symbol(name)])) for name in breakdown_names]
     DuckDB.register_table(connection, breakdown_table, "t_obj_breakdown_solution")
     DuckDB.execute(
         connection,
