### 2026-03-19 6443c86
Add tail excess constraints and tests
