+++ test/test-constraint-tail-excess.jl
@@ -1,0 +1,4 @@
+#= Add a test that checks that the constraints are created when the
+λ
+ parameter is greater than zero (test for one and for more than one scenarios)
+=#
