### 2026-03-19 28bcc95
Add objective cost breakdown to existing components in the objective (#1537)
* Create table and save outputs

Co-Authored-By: Claude Code (claude-sonnet-4-6) <noreply@anthropic.com>

* Refactor to use expressions

Co-Authored-By: Claude Code (claude-sonnet-4-6) <noreply@anthropic.com>

* Apply suggestion from @abelsiqueira

* Fix printing test

* Only print when there is a valid objective value

* Apply suggestions from code review

Co-authored-by: Abel Soares Siqueira <abel.s.siqueira@gmail.com>

* Formatting changes

---------

Co-authored-by: Claude Code (claude-sonnet-4-6) <noreply@anthrop
[…]
