+++ src/objective.jl
@@ -1,3 +1,17 @@
+"""
+    _add_to_objective!(connection, model, objective_expr, name, expr)
+
+Add `expr` to the running objective sum, register it on the model under `name`, and
+insert a placeholder row in the `obj_breakdown` table (value filled in by
+`save_solution!` after the solve).
+"""
+function _add_to_objective!(connection, model, objective_expr, name::String, expr)
+    DuckDB.execute(connection, "INSERT INTO obj_breakdown (name, value) VALUES (?, NULL)", [name])
+    model[Symbol(name)] = expr
+    JuMP.add_to_expression!(objective_expr, expr)
+    return
+end
+
 function add_objective!(connection, model, variables, expressions, profiles, model_parameters)
     assets_investment = variables[:assets_investment]
     assets_investment_energy = variables[:assets_investment_energy]
@@ -21,6 +35,16 @@ function add_objective!(connection, model, variables, expressions, profiles, mod
 
     _create_objective_auxiliary_table(connection, constants)
 
+    ## Create obj_breakdown table (values populated by save_solution! after solve)
+    DuckDB.execute(
+        connection,
+        """CREATE OR REPLACE TABLE obj_breakdown (
+            name  VARCHAR,
+            value FLOAT8
+        )""",
+    )
+    objective_expr = JuMP.AffExpr(0.0)
+
     indices = DuckDB.query(
         connection,
         "SELECT
@@ -43,6 +67,13 @@ function add_objective!(connection, model, variables, expressions, profiles, mod
             row.cost * asset_investment for
             (row, asset_investment) in zip(indices, assets_investment.container)
         )
+    )
+    _add_to_objective!(
+        connection,
+        model,
+        objective_expr,
+        "assets_investment_cost",
+        assets_investment_cost,
     )
 
     # Select expressions for compact method
@@ -72,6 +103,13 @@ function add_objective!(connection, model, variables, expressions, profiles, mod
             zip(indices, expr_available_asset_units_compact_method.expressions[:assets])
         )
     )
+    _add_to_objective!(
+        connection,
+        model,
+        objective_expr,
+        "assets_fixed_cost_compact_method",
+        assets_fixed_cost_compact_method,
+    )
 
     # Select expressions for simple method
     indices = DuckDB.query(
@@ -100,6 +138,13 @@ function add_objective!(connection, model, variables, expressions, profiles, mod
             zip(indices, expr_available_asset_units_simple_method.expressions[:assets])
         )
     )
+    _add_to_objective!(
+        connection,
+        model,
+        objective_expr,
+        "assets_fixed_cost_simple_method",
+        assets_fixed_cost_simple_method,
+    )
 
     indices = DuckDB.query(
         connection,
@@ -123,6 +168,13 @@ function add_objective!(connection, model, variables, expressions, profiles, mod
             row.cost * assets_investment_energy for
             (row, assets_investment_energy) in zip(indices, assets_investment_energy.container)
         )
+    )
+    _add_to_objective!(
+        connection,
+        model,
+        objective_expr,
+        "storage_assets_energy_investment_cost",
+        storage_assets_energy_investment_cost,
     )
 
     indices = DuckDB.query(
@@ -151,6 +203,13 @@ function add_objective!(connection, model, variables, expressions, profiles, mod
             zip(indices, expr_available_energy_units_simple_method.expressions[:energy])
         )
     )
+    _add_to_objective!(
+        connection,
+        model,
+        objective_expr,
+        "storage_assets_energy_fixed_cost",
+        storage_assets_energy_fixed_cost,
+    )
 
     indices = DuckDB.query(
         connection,
[…]
