+++ src/structures.jl
@@ -348,6 +348,18 @@ function Base.show(io::IO, ep::EnergyProblem)
         println(io, "  - Model solved!")
         println(io, "    - Termination status: ", ep.termination_status)
         println(io, "    - Objective value: ", ep.objective_value)
+        if !isnan(ep.objective_value)
+            rows = DuckDB.query(
+                ep.db_connection,
+                "SELECT name, value FROM obj_breakdown ORDER BY name",
+            )
+            if !isempty(rows)
+                println(io, "    - Objective breakdown:")
+                for row in rows
+                    println(io, "      - ", row.name, ": ", row.value)
+                end
+            end
+        end
     elseif !status_model_solved && ep.termination_status == JuMP.INFEASIBLE
         println(io, "  - Model is infeasible!")
     else
