### 2026-03-19 b2d50fc
Add variable Value at Risk threshold mu and tests (#1535)
* Add variable Value at Risk threshold mu and tests

* Update after rebase

* Fix Lint

---------

Co-authored-by: Diego Alejandro Tejada Arango <diego.tejadaarango@tno.nl>
