### 2026-03-20 8a16b91
Make stochastic_scenario table mandatory with explicit CSV files
Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>
