+++ test/utils.jl
@@ -251,5 +251,7 @@ end
             :period => 1:365,
             :rep_period => mod1.(1:365, 2),
         ),
+        "stochastic_scenario" =>
+            DataFrame(:scenario => Int32[1], :probability => [1.0], :description => [""]),
     )
 end
