+++ test/test-bids.jl
@@ -72,6 +72,11 @@
             connection;
             layout = TulipaClustering.ProfilesTableLayout(; year = :milestone_year),
         )
+        DuckDB.query(
+            connection,
+            "CREATE TABLE stochastic_scenario (scenario INTEGER, probability DOUBLE, description VARCHAR)",
+        )
+        DuckDB.query(connection, "INSERT INTO stochastic_scenario VALUES (1, 1.0, '')")
         TulipaEnergyModel.populate_with_defaults!(connection)
 
         return connection
