+++ src/data-preparation.jl
@@ -597,52 +597,3 @@ function _compute_durations(row, max_length_uniform = "")
 
     return error("Row specification '$(row.specification)' is not valid")
 end
-
-"""
-    _calculate_stochastic_scenario_probabilities(connection)
-
-Calculate and populate the stochastic scenario probabilities table.
-
-This function creates a `stochastic_scenario` table with uniform probability distribution
-across all scenarios found in the `rep_periods_mapping` table. If the `stochastic_scenario`
-table already contains data, the function returns early without modifying the table.
-
-# Arguments
-- `connection`: A DuckDB database connection object.
-
-# Behavior
-- If `stochastic_scenario` table is not empty, the function returns `nothing` without making changes.
-- Otherwise, it:
-  1. Extracts a list with the unique values of the scenarios from the `rep_periods_mapping` table
-  2. Calculates uniform probabilities (1/n) for each scenario
-  3. Creates/replaces the `stochastic_scenario` table with scenario names and their probabilities
-
-# Returns
-- `nothing`
-
-# Notes
-- Probabilities are uniformly distributed (each scenario receives probability of 1/number_of_scenarios)
-- Validation of probability values occurs in a separate validation step
-"""
-function _calculate_stochastic_scenario_probabilities(connection)
-    count_scenarios = count_rows_from(connection, "stochastic_scenario")
-    if count_scenarios > 0
-        return nothing
-    end
-
-    scenario_list = DuckDB.query(connection, "SELECT DISTINCT scenario FROM rep_periods_mapping")
-    scenarios = [row.scenario::Int32 for row in scenario_list]
-    scenarios_str = "[" * join(scenarios, ",")::String * "]"
-    num_scenarios = length(scenarios)
-
-    DuckDB.execute(
-        connection,
-        "CREATE OR REPLACE TABLE stochastic_scenario AS
-        SELECT
-            unnest($scenarios_str) AS scenario,
-            1.0 / $num_scenarios AS probability,
-            '' AS description
-        ",
-    )
-    return nothing
-end
