### 2026-03-20 17e3ca9
Increase isapprox tolerance in _is_constraint_equal_kernel
