### 2026-03-20 6927cdb
Fix documentation build failures due to mandatory stochastic_scenario table
Co-authored-by: datejada <12887482+datejada@users.noreply.github.com>
