### 2026-03-20 9baa864
Update docs/src/10-tutorials/40-bids-workaround.md
