### 2026-03-20 94e3e95
WIP: case study stochastic
