+++ test/test-case-studies.jl
@@ -140,7 +140,7 @@ end
     dir = joinpath(INPUT_FOLDER, "TwoStage-StochOpt RPs cross Scenario")
     connection = DBInterface.connect(DuckDB.DB)
     _read_csv_folder(connection, dir)
-    energy_problem = TulipaEnergyModel.run_scenario(connection; show_log = false)
+    energy_problem = TulipaEnergyModel.run_scenario(connection; show_log = true)
     @test energy_problem.objective_value ≈ 10_331_281_729.650423 atol = 1e-5
     # populate_with_defaults shouldn't change the solution
     TulipaEnergyModel.populate_with_defaults!(connection)
@@ -230,4 +230,20 @@ end
     TulipaEnergyModel.populate_with_defaults!(connection)
     energy_problem = TulipaEnergyModel.run_scenario(connection; show_log = false)
     @test energy_problem.objective_value ≈ -404.0 rtol = 1e-8
+end
+
+@testitem "Tiny Case Study where periods are scenarios" setup = [CommonSetup] tags =
+    [:case_study, :integration, :slow] begin
+    dir = joinpath(INPUT_FOLDER, "copy_Tiny")
+    optimizer_list = [HiGHS.Optimizer, GLPK.Optimizer]
+    for optimizer in optimizer_list
+        connection = DBInterface.connect(DuckDB.DB)
+        _read_csv_folder(connection, dir)
+        energy_problem = TulipaEnergyModel.run_scenario(connection; optimizer, show_log = false)
+        @test energy_problem.objective_value ≈ 269238.43825 rtol = 1e-8
+        # populate_with_defaults shouldn't change the solution
+        TulipaEnergyModel.populate_with_defaults!(connection)
+        energy_problem = TulipaEnergyModel.run_scenario(connection; optimizer, show_log = false)
+        @test energy_problem.objective_value ≈ 269238.43825 rtol = 1e-8
+    end
 end
