### 2026-03-21 a898742
implemented ramping with average parameters constraints
