### 2026-03-21 9cb9fd2
add constraints for tight ramping with basic uc
