### 2026-04-02 #9878 @DavidPozo
@marlinarnz I have finished my review. It is a very relevant work for the energy modelling community that deserves wide dissemination.

It follows some comments:

- _AI usage disclosure (required section)_ is not present in the paper, although is is disclosed in the main repository [readme](https://github.com/TulipaEnergy/TulipaEnergyModel.jl?tab=readme-ov-file#ai-coding-assistant-attribution).
- Page 2: “This for direct connection between assets completely avoids…” should be corrected: “This direct connection between assets completely avoids…”
