### 2026-04-07 #9878 @marlinarnz
Thank you @DavidPozo for finishing your review. Does it also include those currently unchecked items in your above list? And do you have any comments to the authors our would you suggest it for publication as is (sorry, if I oversaw issues that are not linked to this thread)?

@trulsf can you give an update on your review timeline?
