### 2026-04-12 TulipaEnergy/TulipaEnergyModel.jl#1578 @trulsf
Missing output directory - Tutorial 1
### Description

While both tutorial 2 and 3 contains a warning about ensuring the output directory existing, this is not mentioned in tutorial 1 where it likely arises first (at least for me). Is it a possibility to automatically create the output directory if it does not exist? E.g. using `mkpath`.
