### 2026-04-12 TulipaEnergy/TulipaEnergyModel.jl#1579 @trulsf
Better structure of API Reference in documentation
### Description

The section with the API Reference in the documentation could benefit with some structure, e.g. dividing into sections with some short intro to each section. The API Reference also includes several internal functions (starting with underscore) that may be left out of the API Reference.
