### 2026-04-12 TulipaEnergy/TulipaEnergyModel.jl#1580 @trulsf
Centralize export declarations
### Description

Has it been considered to move all export declaration to the main package file to get an easier overview of the public API of the package? It is understandable if this is a deliberate choice, but I just wanted to flag it as a possible improvement.
