### 2026-04-12 #9878 @trulsf
Hi,
I have now finished my review of the software package and the paper and I would recommend its acceptance with some minor revisions to the paper. I have raised some issues in the repository on some suggested improvements to the package, but none of these are considered critical for the acceptance (see isssues https://github.com/TulipaEnergy/TulipaEnergyModel.jl/issues/1580 https://github.com/TulipaEnergy/TulipaEnergyModel.jl/issues/1579 and https://github.com/TulipaEnergy/TulipaEnergyModel.jl/issues/1578).

The paper is clearly written and covers most of the JOSS requirements in a good way.
Two minor issues should be addressed before publication:
- There is no section on "AI usage disclosure" and should be added
- While strong on the methodological aspects with several publications, there should be a short section/paragraph on the research impact, describing the use of the software in practical analysis work, preferably with some publications (if possible)
