### 2026-04-13 TulipaEnergy/TulipaEnergyModel.jl#1580 @clizbe
Interesting idea. I see on the EnergyModelsX repo how you've done this. I wonder though if it's more confusing to developers, because in the main file it doesn't seem to show where each exported thing comes from.
It's definitely more clear for what is being exported... 🤔
