### 2026-04-13 TulipaEnergy/TulipaEnergyModel.jl#1578 @clizbe
Great comment! I'll fix it.
If you're running through the tutorials, we really appreciate any and all feedback! :)
