### 2026-04-13 TulipaEnergy/TulipaEnergyModel.jl#1579 @clizbe
@abelsiqueira  What do you think about this? I think at least separating public and internal is a good idea.
But I wonder if there's a way to do this fairly automated - or if this would mean specifying the functions in each section...  (Can we tag them somehow and just filter for by tag for each section?)

If we can automate it, then doing sections beyond just public/internal would also be nice.

Let me know if you have any ideas on how to do this. I can try it in a bit. :)
