### 2026-04-14 TulipaEnergy/TulipaEnergyModel.jl#1579 @abelsiqueira
We can make the public/private split reasonably easily: https://documenter.juliadocs.org/stable/man/syntax/#@autodocs-block

To split more, then you need to manually list the functions, like:
https://github.com/sintefore/TimeStruct.jl/tree/main/docs/src/reference

Manually listing won't be automated, but after the first one is done, we just have to make sure it's not being forgotten. AFAIK, the docs **should** complain if you create a new function with a docstring, and you forgot to manually list it in the `@docs`.
For future reference, maybe we'll need the `canonical` keyword (https://documenter.juliadocs.org/stable/man/syntax/#noncanonical-block)
