### 2026-04-14 TulipaEnergy/TulipaEnergyModel.jl#1580 @datejada
> Indeed, we recently talked about this in [#1516](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/pull/1516), and made the choice deliberate. I don't remember the initial motivation, but it was probably to make it slightly clearer for the developers and less crowded in the main file, though they are not strong reasons. Making the public API more visible was not a strong reason to change, but maybe now that we have so many more functions/structs, and more external packages, it might be better? [@gnawin](https://github.com/gnawin), [@datejada](https://github.com/datejada), opinions?
>
> In the end, what I really miss is a linter tool to validate either option.

Thanks @abelsiqueira for the comments. I don't have a strong opinion, but making the API more visible sounds better.
