### 2026-04-14 TulipaEnergy/TulipaEnergyModel.jl#1580 @abelsiqueira
If @gnawin is also on board, the change is implemented in #1581
