### 2026-04-14 c82d157
Add AI usage disclosure section to JOSS paper and small fix (#1582)
Apparently, we have missed an AI usage disclosure section.
It looks like it was added after we submitted the paper.
