### 2026-04-28 2962491
Add paragraph related to impact to JOSS paper (#1591)
Co-authored-by: Germán Morales-España <german.morales@tno.nl>
