+++ paper/paper.bib
@@ -4,6 +4,14 @@
 
 
 
+@mastersthesis{Kremer2025,
+	title = {Stochastic Programming for Energy Models: A Blended Cross-Scenario Representative Periods Approach},
+	author = {Kremer, L. A. A.},
+	year = {2025},
+	month = jun,
+	school = {Delft University of Technology},
+	url = {https://resolver.tudelft.nl/uuid:0e87f306-1c92-4eec-805f-72377ed57fb2},
+}
 @misc{greg2025,
 	title = {Hull {Clustering} with {Blended} {Representative} {Periods} for {Energy} {System} {Optimization} {Models}},
 	url = {http://arxiv.org/abs/2508.21641},
