### 2026-04-28 #9878 @abelsiqueira
Hi all,

Thanks again, @DavidPozo and @trulsf, for the review. We've updated the paper addressing the comments from the reviewers.

- We've fixed the extra "for" in the phrase mentioned by @DavidPozo ([ref](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/commit/c82d15788fcadc135436e09548656b442db3a158))
- We've added an "AI Usage disclosure" section mentioned by both reviewers ([ref](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/commit/c82d15788fcadc135436e09548656b442db3a158))
- We've added a paragraph with research impact mentioned by @trulsf ([ref](https://github.com/TulipaEnergy/TulipaEnergyModel.jl/commit/2962491f33b7c10b4e5c9a8c92fd0e2cfda43bb4))
- Issue 1580 and 1578, created by @trulsf

We are skipping issue 1579 for now, with the understanding that it is not critical for acceptance.

Please let us know if you identify any more improvements.
Otherwise, let us know the next steps, please, @marlinarnz.
