### 2026-04-30 #9878 @marlinarnz
Thank you for this structured rebuttal. @trulsf and @DavidPozo could you briefly state whether these changes are sufficient to address your comments?
