### 2026-08-20 #9878 @abelsiqueira
Hi @marlinarnz, brief update. We're almost done with the updates. We're in flux as some of the affiliations are changing, and we would like to have up-to-date emails. Hopefully, by the end of the first week of September, we'll have it done.
cf. @datejada @g-moralesespana
